% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAFS.R
\name{AAFS}
\alias{AAFS}
\title{Additive Adaptive Frontier Splines: 1 output}
\usage{
AAFS(
  data,
  x,
  y,
  nterms,
  Kp = 1,
  err.red = 0.01,
  minspan = -2,
  endspan = -2,
  knotsGrid = -1,
  d = 2,
  w1 = seq(1, 2, length.out = 5),
  w2 = seq(8/7, 1.5, length.out = 5),
  na.rm = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{nterms}{For the Forward algorithm. Maximum number of terms created before pruning.}

\item{Kp}{For the Forward algorithm. Maximum degree of interaction allowed. Only \code{1} is available.}

\item{err.red}{For the Forward algorithm. Minimum reduced error rate for the addition of a new pair of basis functions. Default is \code{0.01}.}

\item{minspan}{For the Forward algorithm. Minimum number of observations between two adjacent knots.
\itemize{
\item{\code{minspan = -2}} Computed as in \insertCite{zhang1994;textual}{aafs}.
\item{\code{minspan = -1}} Computed as in \insertCite{friedman1991;textual}{aafs}.
\item{\code{minspan = +m}}
}}

\item{endspan}{For the Forward algorithm. Minimum number of observations before the first and after the final knot.
\itemize{
\item{\code{endspan = -2}} Computed as in \insertCite{zhang1994;textual}{aafs}.
\item{\code{endspan = -1}} Computed as in \insertCite{friedman1991;textual}{aafs}.
\item{\code{endspan = +m}}
}}

\item{knotsGrid}{For the Forward algorithm. If \code{knotsGrid == -1} (default), the original approach of \insertCite{friedman1991;textual}{aafs} based on the observed data is used. Alternatively, a \code{list} with the grid of knots to perform AAFS can be entered. In this case, each element of the \code{list} contains a vector with the knots of one variable (e.g., the second element of the list contains the knot values of the second variable).}

\item{d}{For the Backward algorithm. Generalized Cross Validation (GCV) penalty per knot. Default is \code{2}.}

\item{w1}{For the cubic smoothing procedure \insertCite{friedman1991}{aafs}. The distance between the central knot and a side knot is at most \code{w} times the distance between that central knot and the other side knot for each truncated cubic basis function. It must be set between 1 and 2. If a \code{vector} is entered, the \code{w} value that most reduce the lack-of-fit is selected.}

\item{w2}{For the quintic smoothing procedure \insertCite{chen1999}{aafs}. The distance between the central knot and a side knot is at most \code{w} times the distance between that central knot and the other side knot for each truncated quintic basis function. It must be set between 8/7 and 1.5. If a \code{vector} is entered, the \code{w} value that most reduce the lack-of-fit is selected.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
An \code{AAFS} object.
}
\description{
This function estimates a production frontier satisfying some classical production theory axioms, such as monotonicity and concavity, which is based upon the adaptation of the machine learning technique known as Multivariate Adaptive Regression Splines (MARS) developed by \insertCite{friedman1991;textual}{aafs}.
}
\references{
\insertRef{zhang1994}{aafs} \cr
\cr
\insertRef{friedman1991}{aafs} \cr
\cr
\insertRef{chen1999}{aafs}
}
