% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{AAFS_BCC.OUT}
\alias{AAFS_BCC.OUT}
\title{The output-oriented Banker, Charnes and Cooper Programming Model for Additive Adaptive Frontier Splines}
\usage{
AAFS_BCC.OUT(dmu, xmat, ymat, aknot, yknot, nX, nY, nk)
}
\arguments{
\item{dmu}{Number of DMUs in the sample.}

\item{xmat}{\code{data.frame} or \code{matrix} containing the inputs in the model.}

\item{ymat}{\code{data.frame} or \code{matrix} containing the outputs in the model.}

\item{aknot}{Knots coordinates.}

\item{yknot}{Predictions for knots.}

\item{nX}{Number of inputs.}

\item{nY}{Number of outputs.}

\item{nk}{Number of knots.}
}
\value{
A numerical vector with the efficiency scores computed.
}
\description{
This function computes efficiency scores through a BCC model with output orientation for AAFS.
}
