% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAFS.R
\name{AAFS_object}
\alias{AAFS_object}
\title{Create an AAFS object}
\usage{
AAFS_object(
  data,
  x,
  y,
  nterms,
  Kp,
  err.red,
  minspan,
  endspan,
  knotsGrid,
  d,
  na.rm,
  Forward.AAFS,
  AAFS,
  QSmooth.AAFS,
  CSmooth.AAFS
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{nterms}{Maximum number of terms created by the forward algorithm (before pruning).}

\item{Kp}{Maximum degree of interaction allowed. Only \code{1} is available.}

\item{err.red}{Minimum reduced error rate for the addition of two new basis functions.}

\item{minspan}{Minimum number of observations between two adjacent knots.}

\item{endspan}{Minimum number of observations before the first and after the final knot.}

\item{knotsGrid}{Grid of knots to perform AAFS.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.}

\item{Forward.AAFS}{\code{list} containing the Forward Additive Adaptive Frontier Splines.}

\item{AAFS}{\code{list} containing the Additive Adaptive Frontier Splines model.}

\item{QSmooth.AAFS}{\code{list} containing the Smooth Additive Adaptive Frontier Splines model (quintic version).}

\item{CSmooth.AAFS}{\code{list} containing the Smooth Additive Adaptive Frontier Splines model (cubic version).}
}
\value{
A \code{AAFS} object.
}
\description{
This function saves information about the Additive Adaptive Frontier Splines model.
}
