% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_scores.R
\name{aces_scores}
\alias{aces_scores}
\title{Compute Efficiency Scores using an Adaptive Constrained Enveloping Splines model.}
\usage{
aces_scores(
  tech_data,
  eval_data,
  x,
  y,
  addi_x,
  object,
  method,
  proximity = ifelse(length(y) == 1, Inf, 0.05),
  measure = "rad_out",
  convexity = TRUE,
  returns = "variable",
  direction = NULL,
  weights = NULL,
  digits = 3
)
}
\arguments{
\item{tech_data}{A \code{data.frame} or a \code{matrix} containing the observed DMUs to determine the technology.}

\item{eval_data}{A \code{data.frame} or a \code{matrix} containing the DMUs to be evaluated.}

\item{x}{Column indexes of input variables in \code{tech_data} and \code{eval_data}.}

\item{y}{Column indexes of output and netput variables in \code{tech_data} and \code{eval_data}.}

\item{addi_x}{Column indexes of additional covariables in \code{data}.}

\item{object}{An \code{aces} object.}

\item{method}{Model prediction used to compute predictions of inputs and obtain a new vector of outputs for \code{eval_data} .
\itemize{
\item{\code{"aces_forward"}}: Forward Adaptive Constrained Enveloping Splines model.
\item{\code{"aces"}}: Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_cubic"}}: Cubic Smoothed Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_quintic"}}: Quintic Smoothed Adaptive Constrained Enveloping Splines model.
}}

\item{proximity}{A value indicating the proximity of the DMUs to the DEA-VRS frontier.}

\item{measure}{Mathematical programming model to calculate scores:
\itemize{
\item{\code{rad_out}} The output-oriented radial measure. Efficiency level at 1. (default)
\item{\code{rad_inp}} The input-oriented radial measure. Efficiency level at 1.
\item{\code{ddf}}     The directional distance function. Efficiency level at 0.
\item{\code{rsl_out}} The output-oriented Russell measure. Efficiency level at 1.
\item{\code{rsl_inp}} The input-oriented Russell measure. Efficiency level at 1.
\item{\code{wam_mip}} The weighted additive model: measure of inefficiency proportions. Efficiency level at 0.
\item{\code{wam_ram}} The weighted additive model: range adjusted measure of inefficiency. Efficiency level at 0.
}}

\item{convexity}{A \code{logical} value indicating if a convex technology is assumed. Default is \code{TRUE}.}

\item{returns}{Type of returns to scale:
\itemize{
\item{\code{"variable"}} Variable returns to scale (default).
\item{\code{"constant"}} Constant returns to scale.
}}

\item{direction}{Only applied if \code{measure = "ddf"}. Direction of the vector to project on the frontier:
\itemize{
\item{\code{"mean"}} Projection vector given by the average value of inputs and outputs of all DMUs. Applied if \code{direction = NULL}.
\item{\code{"briec"}} Projection vector given by the value of inputs and outputs of the evaluated DMU.
}}

\item{weights}{Weights for the additive model:
\itemize{
\item{\code{"WAM"}} Weighted Additive Model.
\item{\code{"MIP"}} Measure of Inefficiency Proportions.
\item{\code{"NOR"}} Normalized Weighted Additive Model.
\item{\code{"RAM"}} Range Adjusted Measure.
\item{\code{"BAM"}} Bounded Adjusted Measure.
}}

\item{digits}{Number of decimal places to which efficiency scores are rounded.}
}
\value{
A \code{data.frame} with the efficiency scores computed through an Adaptive Constrained Enveloping Splines model.
}
\description{
This function calculates the efficiency scores for each Decision-Making-Unit through an Adaptive Constrained Enveloping Splines model and a specific technology determined by the user.
}
