% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acesFunctions.R
\name{comparison_aces}
\alias{comparison_aces}
\title{Apply ACES (Automatic Cartridge Evidence Scoring) comparison algorithm to a
set of cartridge case scans.}
\usage{
comparison_aces(
  ...,
  x3p_labels = NULL,
  return_registrations = FALSE,
  standardize_resolutions = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{...}{either a sequence of x3p objects separated by commas like
'x3p1,x3p2,x3p3,...', a single list of x3p objects separated by commas like
'list(x3p1,x3p2,x3p3,...)', or a vector of x3p file paths like
'c("path/to/file1.x3p", "path/to/file2.x3p", "path/to/file3.x3p")'}

\item{x3p_labels}{(optional) character vector of x3p labels}

\item{return_registrations}{whether to return the full scan and cell-based
registrations as tibble columns, each containing nested tibble objects.
This is useful for understanding the feature values, but greatly increases
the size of the returned object.}

\item{standardize_resolutions}{whether to force resolutions of all scans
equal. If FALSE, then function will throw error when scans aren't same
resolution.}

\item{quiet}{whether to print internal function messages to the console}
}
\description{
Apply ACES (Automatic Cartridge Evidence Scoring) comparison algorithm to a
set of cartridge case scans.
}
