% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{add_scenario_XnY1}
\alias{add_scenario_XnY1}
\title{Simulate an Additive Production Scenario}
\usage{
add_scenario_XnY1(N, scenario)
}
\arguments{
\item{N}{Positive integer giving the number of observations.}

\item{scenario}{Scenario to simulate: \code{"A"}, \code{"B"}, \code{"C"}, \code{"D"},
\code{"E"}, or \code{"F"}.}
}
\value{
A data frame with one to three inputs, observed output \code{y}, and frontier
output \code{yT}.
}
\description{
Simulates one of six additive production scenarios from
\insertCite{kuosmanen2010;textual}{aces}.
}
\references{
\insertRef{kuosmanen2010}{aces}
}
