% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardAlgorithm.R
\name{AddBF}
\alias{AddBF}
\title{Add a New Pair of Basis Functions}
\usage{
AddBF(
  data,
  x,
  y,
  ForwardModel,
  knots.list,
  Kp,
  L,
  Le,
  knotsGrid,
  linpreds,
  err.min
)
}
\arguments{
\item{data}{data \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{ForwardModel}{\code{list} containing the set of basis functions and the B matrix.}

\item{knots.list}{\code{list} containing the set of selected knots.}

\item{Kp}{Maximum degree of interaction allowed.}

\item{L}{Minimum number of observations between two adjacent knots.}

\item{Le}{Minimum number of observations before the first and after the final knot.}

\item{knotsGrid}{Grid of virtual knots to perform MAFS. Can be set to \code{NULL}.}

\item{linpreds}{\code{logical}. If \code{TRUE}, predictors can enter linearly.}

\item{err.min}{Minimun error in the split.}
}
\value{
A \code{list} containing the matrix of basis functions (\code{B}), a \code{list} of basis functions (\code{BF}), a \code{list} of selected knots (\code{knots.list}) and the minimun error (\code{err.min}).
}
\description{
This function adds the best pair of basis functions to the model.
}
