% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{basis_to_drop}
\alias{basis_to_drop}
\title{Find Removable Basis Functions}
\usage{
basis_to_drop(Bp_list, conc, nX)
}
\arguments{
\item{Bp_list}{Basis functions grouped by input.}

\item{conc}{If \code{TRUE}, preserve concavity when selecting candidates.}

\item{nX}{Number of inputs.}
}
\value{
A data frame of candidate basis functions to remove.
}
\description{
Finds the basis functions that can be removed from each input.
}
