% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2NLS.R
\name{C2NLS}
\alias{C2NLS}
\title{Corrected Concave Nonparametric Least Squares}
\usage{
C2NLS(data, x, y)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{y}{output indexes in \code{data}.}

\item{X}{input indexes in \code{data}.}
}
\value{
A \code{C2NLS} object.
}
\description{
This function performs Corrected Concave Nonparametric Least Squares.
}
