% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2NLS.R
\name{C2NLS_object}
\alias{C2NLS_object}
\title{Create a C2NLS object}
\usage{
C2NLS_object(data, x, y, alpha, beta, epsilon, f)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{alpha}{alpha variable.}

\item{beta}{beta variable.}

\item{epsilon}{epsilon variable.}

\item{f}{alpha + beta * x + e.}
}
\value{
A \code{C2NLS} object.
}
\description{
This function saves information about the Corrected Concave Nonparametric Least Squares model.
}
