% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2NLS.R
\name{CNLS}
\alias{CNLS}
\title{Corrected Nonparametric Least Squares}
\usage{
CNLS(data, x, y)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{y}{output indexes in \code{data}.}

\item{X}{input indexes in \code{data}.}
}
\value{
\code{data.frame} with the solutions of the optimization problem solved.
}
\description{
This function performs Corrected Nonparametric Least Squares.
}
