% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cobb_douglas_XnY1}
\alias{cobb_douglas_XnY1}
\title{Simulate Cobb-Douglas Data}
\usage{
cobb_douglas_XnY1(N, nX)
}
\arguments{
\item{N}{Positive integer giving the number of observations.}

\item{nX}{Number of inputs. Allowed values are \code{1}, \code{3}, \code{6}, \code{9},
\code{12}, and \code{15}.}
}
\value{
A data frame with inputs \code{x1}, ..., observed output \code{y}, and frontier
output \code{yT}.
}
\description{
Simulates inputs, one inefficient output, and its Cobb-Douglas frontier value.
}
