% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cobb_douglas_XnY1_noise}
\alias{cobb_douglas_XnY1_noise}
\title{Simulate Noisy Cobb-Douglas Data}
\usage{
cobb_douglas_XnY1_noise(N, nX, p, heteroskedasticity = FALSE)
}
\arguments{
\item{N}{Positive integer giving the number of observations.}

\item{nX}{Number of inputs. Allowed values are \code{1} and \code{3}.}

\item{p}{Non-negative noise-to-signal ratio.}

\item{heteroskedasticity}{If \code{TRUE}, use heteroskedastic inefficiency and noise. This is available
only when \code{nX = 3}.}
}
\value{
A data frame with inputs \code{x1}, ..., observed output \code{y}, and frontier
output \code{yT}.
}
\description{
Simulates inputs, one inefficient output with statistical noise, and its
Cobb-Douglas frontier value.
}
\references{
\insertRef{simar2011}{aces}
}
