% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_ACES.R
\name{compute_variable_importance}
\alias{compute_variable_importance}
\title{Compute Variable Importance in Random Forest Adaptive Constrained Enveloping Splines (RF-ACES).}
\usage{
compute_variable_importance(data, x, y, object, repeats = 1)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{object}{A \code{rf_aces} object.}

\item{repeats}{Number of times the variable importance procedure is repeated.}
}
\value{
This function returns a metric of variable importance for each input variable
}
\description{
This function computes a metric of variable importance for a given Random Forest Adaptive Constrained Enveloping Splines (RF-ACES) with a certain number of base learners.
}
