% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackwardAlgorithm.R
\name{computeGCV}
\alias{computeGCV}
\title{Compute Generalized Cross-Validation}
\usage{
computeGCV(data, y, B, coefs, d, knots)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{y}{Column output index in \code{data}.}

\item{B}{\code{matrix} of basis functions.}

\item{coefs}{\code{vector} of coefficients in the model.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{knots}{\code{list} of selected knots.}
}
\value{
Generalized Cross-Validation value
}
\description{
This function computes the generalized cross-validation for the backward procedure.
}
