% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAFS.R
\name{computeSpan}
\alias{computeSpan}
\title{Compute Span}
\usage{
computeSpan(data, minspan, endspan, nX)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{minspan}{Minimum number of observations between two adjacent knots.
\itemize{
\item{\code{minspan = -2}} Computed as in \insertCite{zhang1994;textual}{aafs}.
\item{\code{minspan = -1}} Computed as in \insertCite{friedman1991;textual}{aafs}.
\item{\code{minspan = +m}}
}}

\item{endspan}{Minimum number of observations before the first and after the final knot.
\itemize{
\item{\code{endspan = -2}} Computed as in \insertCite{zhang1994;textual}{aafs}.
\item{\code{endspan = -1}} Computed as in \insertCite{friedman1991;textual}{aafs}.
\item{\code{endspan = +m}}
}}

\item{nX}{Number of inputs.}
}
\value{
\code{list} with the minimum span and the end span.
}
\description{
This function computes the minimum span and the end span.
}
