% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients.R
\name{concavity_matrix}
\alias{concavity_matrix}
\title{Build the Concavity Constraint Matrix}
\usage{
concavity_matrix(Bp_list)
}
\arguments{
\item{Bp_list}{Basis functions grouped by input.}
}
\value{
A constraint matrix for the basis coefficients.
}
\description{
Builds the constraints that enforce concavity in the reduced linear program.
}
