% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_ACES.R
\name{cond_eff}
\alias{cond_eff}
\title{Conditional Efficiency Estimation}
\usage{
cond_eff(data, y, mean_pred, std_ineff, std_error, error_type)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{mean_pred}{A \code{vector} of mean predicted outputs.}

\item{std_ineff}{A \code{numeric} indicating the Standard Deviation of the Inefficiency Term.}

\item{std_error}{A \code{numeric} indicating the Standard Deviation of the Error Term.}

\item{error_type}{A \code{character} string specifying the error structure when fitting the model.}
}
\value{
A \code{vector} of \code{"numeric"} scores computed through the Conditional Estimation.
}
\description{
This function computes efficiency scores assuming a zero-truncated normal distribution.
}
