% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_algorithm.R
\name{create_linear_basis}
\alias{create_linear_basis}
\title{Create Linear Basis Functions}
\usage{
create_linear_basis(data, var_idx, knot)
}
\arguments{
\item{data}{A matrix containing the prepared model variables.}

\item{var_idx}{Index of the input used to create the basis functions.}

\item{knot}{Knot location.}
}
\value{
A list containing the right and left hinge vectors.
}
\description{
Creates the two hinge functions on either side of a knot.
}
