% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth.R
\name{CreateCubicBF}
\alias{CreateCubicBF}
\title{Generate a new pair of Cubic Basis Functions}
\usage{
CreateCubicBF(data, xi, knt, B, side, wl)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{xi}{Variable index of the new basis function(s).}

\item{knt}{Knots for creating the new basis function(s).}

\item{B}{Matrix of basis functions.}

\item{side}{Side of the basis function.}

\item{wl}{The distance between the central knot and a side knot is at most \code{wl} times the distance between that central knot and the other side knot. It must be between 1 and 2.}
}
\value{
Matrix of basis functions updated with the new basis functions.
}
\description{
This function generates two new cubic basis functions from a variable and a knot previously created during AAFS algorithm.
}
