% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth.R
\name{CSAAFS}
\alias{CSAAFS}
\title{Smoothing Additive Adaptive Frontier Splines through Cubic Functions}
\usage{
CSAAFS(data, nX, knots, x.space, y, w)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{nX}{number of inputs in \code{data}.}

\item{knots}{\code{data.frame} containing the knots from Backward AAFS.}

\item{x.space}{\code{list} with the input space of each variable for smoothing.}

\item{y}{output indexes in \code{data}.}

\item{w}{The distance between the central knot and a side knot is at most \code{w} times the distance between that central knot and the other side knot for each truncated cubic basis function.}
}
\value{
List containing the set of knots from backward (\code{knots}), the new cubic knots (\code{cubic_knots}) and the set of coefficients (\code{coefs}).
}
\description{
This function smoothes the AAFS predictor and imposes continuous first derivatives.
}
