% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DEA_BCC_out}
\alias{DEA_BCC_out}
\title{Model prediction for DEA}
\usage{
DEA_BCC_out(j, scores, x_k, y_k, atreeTk, ytreeTk, nX, nY, N_leaves)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}
}
\value{
\code{data.frame} with the original data and the predicted values through a DEA model.
}
\description{
This function predicts the expected output through a DEA model.
}
