% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{display_errors}
\alias{display_errors}
\title{Error Messaging in ACES functions.}
\usage{
display_errors(
  data,
  x,
  y,
  y_type,
  addi_x,
  degree,
  metric,
  knots_grid,
  d,
  wc,
  wq,
  object,
  measure,
  returns,
  direction,
  digits
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{y_type}{The mode of predicting the output: whether it should be done all at once or one at a time.}

\item{addi_x}{Column indexes of additional covariables in \code{data}.}

\item{degree}{Maximum degree of interaction between variables.}

\item{metric}{The lack-of-fit criterion to evaluate the model performance.}

\item{knots_grid}{Grid of knots to perform ACES.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{wc}{Hyperparameter for the side knot distances in the cubic smoothing procedure \insertCite{friedman1991}{aces}.}

\item{wq}{Hyperparameter for the side knot distances in the quintic smoothing procedure \insertCite{chen1999}{aces}.}

\item{object}{An \code{aces} object.}

\item{measure}{Mathematical programming model to compute the efficiency scores.}

\item{returns}{Type of returns to scale for computing the efficiency scores.}

\item{direction}{Vector direction for DMU projection in Directional Distance Function when computing the efficiency scores.}

\item{digits}{Number of digits to round efficiency scores.}
}
\value{
Possible error messages.
}
\description{
This function displays error messages if hyperparameters or data are bad introduced.
}
\references{
\insertRef{friedman1991}{aces} \cr
\cr
\insertRef{chen1999}{aces}
}
