% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_linear.R
\name{ec_envelopment}
\alias{ec_envelopment}
\title{Estimate Coefficients in Adaptive Constrained Enveloping Splines Fitting: Envelopment Version}
\usage{
ec_envelopment(
  B,
  y_obs,
  dea_eff,
  it_list,
  Bp_list,
  monotonicity,
  concavity,
  origin
)
}
\arguments{
\item{B}{A \code{matrix} of linear basis functions.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{dea_eff}{An indicator vector with 1s for efficient DMUs and 0s for inefficient DMUs.}

\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{monotonicity}{A \code{logical} value indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} value indicating whether to enforce the constraint of concavity in the estimator.}

\item{origin}{A \code{logical} value indicating whether the estimator should satisfy f(0) = 0.}
}
\value{
A \code{vector} of estimated coefficients.
}
\description{
This function solves a Linear Programming problem to obtain a vector of coefficients that impose the required shaped on the Adaptive Constrained Enveloping Splines estimator in the envelopment version.
}
