% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_linear.R
\name{ec_stochastic}
\alias{ec_stochastic}
\title{Estimate Coefficients in Adaptive Constrained Enveloping Splines Fitting: Stochastic Version}
\usage{
ec_stochastic(B, y_obs, it_list, Bp_list, monotonicity, concavity, origin)
}
\arguments{
\item{B}{A \code{matrix} of linear basis functions.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{monotonicity}{A \code{logical} value indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} value indicating whether to enforce the constraint of concavity in the estimator.}

\item{origin}{A \code{logical} value indicating whether the estimator should satisfy f(0) = 0.}
}
\value{
A \code{vector} of estimated coefficients.
}
\description{
This function solves a Quadratic Programming problem to obtain a vector of coefficients that impose the required shaped on the Adaptive Constrained Enveloping Splines estimator in the stochastic version
}
