% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_smooth.R
\name{ecs_stochastic}
\alias{ecs_stochastic}
\title{Estimate Coefficients in Smoothing Adaptive Constrained Enveloping Splines Fitting: Stochastic Version and Additive Error}
\usage{
ecs_stochastic(
  B,
  y_obs,
  dea_eff,
  n_pair,
  n_lsub,
  monotonicity,
  concavity,
  origin
)
}
\arguments{
\item{B}{A \code{matrix} of cubic or quintic basis functions.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{n_pair}{An \code{integer} specifying the number of coefficients associated with the paired basis functions.}

\item{n_lsub}{An \code{integer} specifying the number of coefficients associated with the unpaired left-side basis functions.}

\item{monotonicity}{A \code{logical} value indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} value indicating whether to enforce the constraint of concavity in the estimator.}

\item{origin}{A \code{logical} value indicating whether the estimator should satisfy f(0) = 0.}
}
\value{
A \code{vector} of estimated coefficients.
}
\description{
This function solves a Linear Programming problem to obtain a vector of coefficients that impose the required shaped on the Smoothing Adaptive Constrained Enveloping Splines estimator in the stochastic version, assuming an additive error structure.
}
