% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_smooth.R
\name{estim_coefs_smooth_envelopment}
\alias{estim_coefs_smooth_envelopment}
\title{Estimate Coefficients in Smooth Adaptive Constrained Enveloping Splines Fitting: Envelopment Version and Additive Error}
\usage{
estim_coefs_smooth_envelopment(B, y_obs, dea_scores, n_pair, n_lsub, shape)
}
\arguments{
\item{B}{A \code{matrix} of cubic or quintic basis functions.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{dea_scores}{A \code{matrix} containing DEA-VRS efficiency scores, calculated using an output-oriented radial model. For models with multiple outputs, each column corresponds to the scores for one specific output.}

\item{n_pair}{An \code{integer} specifying the number of coefficients associated with the paired basis functions.}

\item{n_lsub}{An \code{integer} specifying the number of coefficients associated with the unpaired left-side basis functions.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity.}
}
\value{
A \code{vector} of estimated coefficients.
}
\description{
This function solves a Linear Programming problem to obtain a vector of coefficients that impose the required shaped on the Smooth Adaptive Constrained Enveloping Splines estimator in the envelopment version, assuming an additive error structure.
}
