% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_linear.R
\name{estimate_coefficients_envelopment_contextual}
\alias{estimate_coefficients_envelopment_contextual}
\title{Estimate Coefficients in Adaptive Constrained Enveloping Splines: Envelopment Version with Contextual Variables}
\usage{
estimate_coefficients_envelopment_contextual(
  B,
  Z,
  y_obs,
  dea_scores,
  it_list,
  Bp_list,
  shape
)
}
\arguments{
\item{B}{A \code{matrix} of linear basis functions derived from input variables.}

\item{Z}{A \code{matrix} of linear basis functions derived from contextual variables.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{dea_scores}{A \code{matrix} containing DEA-VRS efficiency scores, calculated using an output-oriented radial model. For models with multiple outputs, each column corresponds to the scores for one specific output.}

\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity.}
}
\value{
A \code{vector} containing the estimated coefficients for the ACES model.
}
\description{
This function estimates a vector of coefficients for the Adaptive Constrained Enveloping Splines (ACES) model in its envelopment version using contextual variables. It solves a mathematical programming problem to determine the coefficients that best fit the observed data while respecting certain shape constraints such as monotonicity and/or concavity. The model aims to envelop the data points, particularly when estimating production functions or efficiency frontiers.
}
