% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_linear.R
\name{estimate_coefficients_stochastic}
\alias{estimate_coefficients_stochastic}
\title{Estimate Coefficients in Adaptive Constrained Enveloping Splines Fitting: Stochastic Version}
\usage{
estimate_coefficients_stochastic(B, y_obs, it_list, Bp_list, shape)
}
\arguments{
\item{B}{A \code{matrix} of linear basis functions derived from input variables.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity.}
}
\value{
A \code{vector} containing the estimated coefficients for the ACES model.
}
\description{
This function estimates a vector of coefficients for the Adaptive Constrained Enveloping Splines (ACES) model in its stochastic version. It solves a quadratic programming problem to determine the coefficients that best fit the observed data while respecting certain shape constraints such as monotonicity and/or concavity.
}
