% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{homothetic_s_shaped}
\alias{homothetic_s_shaped}
\title{1 output ~ 2 inputs S-shaped homothetic production function}
\usage{
homothetic_s_shaped(N, noise)
}
\arguments{
\item{N}{Sample size.}

\item{noise}{Random noise. 4 possible levels: \code{0.01}, \code{0.05}, \code{0.1} or \code{0.15}.}
}
\value{
A \code{data.frame} with the simulated data: 2 inputs (x1, x2, x3), 1 output (y1) and the theoretical frontier (yD)
}
\description{
This function is used to simulate a 1 output ~ 2 inputs \code{data.frame} with a homothetic S-shaped Data Generation Process as in \insertCite{olesen2022;textual}{aces}.
}
\references{
\insertRef{olesen2022}{aces}
}
\keyword{internal}
