% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackwardAlgorithm.R
\name{InteractionModel}
\alias{InteractionModel}
\title{Add interaction of variables}
\usage{
InteractionModel(data, x, y, model, knots, err.red, Kp, d)
}
\arguments{
\item{data}{\code{Data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{model}{MAFS model from backward algorithm.}

\item{knots}{Knots from (backward) MAFS algorithm.}

\item{err.red}{Minimum reduced error rate for the addition of two new basis functions. Default is \code{0.01}.}

\item{Kp}{Maximum degree of interaction allowed. Default is \code{1}.}

\item{d}{Integer. Generalized Cross Validation (GCV) penalty per knot. If set to \code{-1}, \code{GCV = RSS / n}.}
}
\value{
\code{vector} with the coefficients estimated.
}
\description{
This function adds interaction of variables in the model.
}
