% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{MAFS.scores}
\alias{MAFS.scores}
\title{Efficiency Scores computed through a Multivariate Adaptive Frontier Splines model.}
\usage{
MAFS.scores(
  data,
  x,
  y,
  object,
  scores_model,
  digits = 3,
  DEA = TRUE,
  print.table = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{object}{A \code{MAFS} object.}

\item{scores_model}{Mathematical programming model to calculate scores.
\itemize{
\item{\code{BCC.OUT}} BCC model. Output-oriented. Efficiency level at 1.
\item{\code{BCC.INP}} BCC model. Input-oriented. Efficiency level at 1.
\item{\code{DDF}}     Directional Distance Function. Efficiency level at 0.
\item{\code{RSL.OUT}} Russell model. Output-oriented. Efficiency level at 1.
\item{\code{RSL.INP}} Russell model. Input-oriented. Efficiency level at 1.
\item{\code{WAM.MIP}} Weighted Additive Model. Measure of Inefficiency Proportions. Efficiency level at 0.
\item{\code{WAM.RAM}} Weighted Additive Model. Range Adjusted Measure of Inefficiency. Efficiency level at 0.
}}

\item{digits}{Decimal units for scores.}

\item{DEA}{\code{logical}. If \code{TRUE}, DEA scores are also computed with the programming model selected in \code{scores_model}.}

\item{print.table}{\code{logical}. If \code{TRUE}, a summary descriptive table of the efficiency scores is displayed.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
A \code{data.frame} with the efficiency scores computed through a Multivariate Adaptive Frontier Splines model.
}
\description{
This function computes the efficiency scores for each DMU through a Multivariate Adaptive Frontier Splines model.
}
