% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{MAFS_RSL.OUT}
\alias{MAFS_RSL.OUT}
\title{The output-oriented Russell Model for Multivariate Adaptive Frontier Splines}
\usage{
MAFS_RSL.OUT(dmu, xmat, ymat, aknot, yknot, nX, nY, nk)
}
\arguments{
\item{dmu}{Number of DMUs in the smample.}

\item{xmat}{\code{data.frame} or \code{matrix} containing the inputs in the model.}

\item{ymat}{\code{data.frame} or \code{matrix} containing the outputs in the model.}

\item{aknot}{Knots coordinates.}

\item{yknot}{Predictions for knots.}

\item{nX}{Number of inputs.}

\item{nY}{Number of outputs.}

\item{nk}{Number of knots.}
}
\value{
A numerical vector with efficiency scores.
}
\description{
This function computes efficiency scores through a Russell model with output orientation for MAFS.
}
