% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scores.R
\name{MAFS_WAM}
\alias{MAFS_WAM}
\title{The Weighted Additive Model for Multivariate Adaptive Frontier Splines}
\usage{
MAFS_WAM(dmu, xmat, ymat, aknot, yknot, nX, nY, nk, weights)
}
\arguments{
\item{dmu}{Number of DMUs in the smample.}

\item{xmat}{\code{data.frame} or \code{matrix} containing the inputs in the model.}

\item{ymat}{\code{data.frame} or \code{matrix} containing the outputs in the model.}

\item{aknot}{Knots coordinates.}

\item{yknot}{Predictions for knots.}

\item{nX}{Number of inputs.}

\item{nY}{Number of outputs.}

\item{nk}{Number of knots.}

\item{weights}{Character. \code{"MIP"} for Measure of Inefficiency Proportion or \code{"RAM"} for Range Adjusted Measure of Inefficiency.}
}
\value{
A numerical vector with efficiency scores.
}
\description{
This function computes efficiency scores through a WAM model for MAFS.
}
