% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardAlgorithm.R
\name{me}
\alias{me}
\title{Mean Error}
\usage{
me(y, y_pred)
}
\arguments{
\item{y}{Vector of actual data.}

\item{y_pred}{Vector of predicted values.}
}
\value{
Mean Error.
}
\description{
This function computes the mean error between two numeric vectors.
}
