% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{MultScenario}
\alias{MultScenario}
\title{Multiplicative Scenarios for 1 Output}
\usage{
MultScenario(N, scenario)
}
\arguments{
\item{N}{Sample size.}

\item{scenario}{\code{"A"}, \code{"B"}, \code{"C"}, \code{"D"}, \code{"E"} or \code{"F"}. For details, check Table 4.}
}
\value{
\code{data.frame} with the simulated data.

\insertRef{kuosmanen2010}{mafs}
}
\description{
This function is used to simulate the data in different multiplicative scenarios as in \insertCite{kuosmanen2010;textual}{mafs}.
}
