% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.AAFS}
\alias{predict.AAFS}
\title{Model Prediction for Additive Adaptive Frontier Splines.}
\usage{
\method{predict}{AAFS}(object, newdata, x, class = 2)
}
\arguments{
\item{object}{A \code{AAFS} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{class}{Model for prediction. \code{1} for AAFS before pruning, \code{2} for AAFS, \code{3} for CSAAFS (cubic version) and \code{4} for QSAAFS (quintic version) and .}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by a \code{AAFS} object.
}
