% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2NLS.R
\name{predict.C2NLS}
\alias{predict.C2NLS}
\title{Model Prediction for Corrected Concave Nonparametric Least Squares}
\usage{
\method{predict}{C2NLS}(object, newdata, x)
}
\arguments{
\item{object}{A \code{C2NLS} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by a \code{C2NLS} object.
}
