% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{predict.MAFS}
\alias{predict.MAFS}
\title{Model Prediction for Multivariate Adaptive Frontier Splines.}
\usage{
\method{predict}{MAFS}(object, newdata, x, class = 2)
}
\arguments{
\item{object}{A \code{MAFS} object.}

\item{newdata}{\code{data.frame}. Set of input variables to predict on.}

\item{x}{Inputs index.}

\item{class}{Model for prediction. \code{1} for MAFS before pruning, \code{2} for MAFS, \code{3} for Smooth MAFS and \code{4} for Smooth Forward MAFS.}
}
\value{
\code{data.frame} with the predicted values.
}
\description{
This function predicts the expected output by a \code{MAFS} object.
}
