% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.rf_aces}
\alias{predict.rf_aces}
\title{Predict Method for RF-ACES Models}
\usage{
\method{predict}{rf_aces}(object, newdata, x, method = "rf_aces", ...)
}
\arguments{
\item{object}{A \code{rf_aces} object.}

\item{newdata}{A data frame or matrix containing the new observations.}

\item{x}{Column indexes of input variables in \code{newdata}.}

\item{method}{RF-ACES model to use.}

\item{...}{Additional arguments. They are currently ignored.}
}
\value{
A data frame of predicted outputs.
}
\description{
S3 prediction method for \code{rf_aces} objects. It delegates to
\code{rf_aces_predict()} and returns the ensemble output predictions.
}
