% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.s_aces}
\alias{predict.s_aces}
\title{Model Prediction for Adaptive Constrained Enveloping Splines (ACES).}
\usage{
\method{predict}{s_aces}(object, newdata, x, method = "aces", stochastic_pred = FALSE)
}
\arguments{
\item{object}{An \code{aces} object.}

\item{newdata}{A \code{data.frame} containing the input and netput variables to predict on.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{method}{Model for prediction:
\itemize{
\item{\code{"aces_forward"}}: Forward Adaptive Constrained Enveloping Splines model.
\item{\code{"aces"}}: Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_cubic"}}: Cubic Smoothed Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_quintic"}}: Quintic Smoothed Adaptive Constrained Enveloping Splines model.
}}

\item{stochastic_pred}{This parameter should be kept as \code{FALSE} (default) if an additive model is estimated. If a stochastic model is estimated, one of the following elements should be specified:
\itemize{
  \item{"avg"}: average-practice production function.
  \item{"mom"}: moments-estimation production function.
  \item{"pse"}: pseudolikelihood-estimation production function.
}}
}
\value{
A \code{data.frame} with the predicted values through the Adaptive Constrained Enveloping Splines model.
}
\description{
This function predicts the expected output by an \code{aces} object.
}
