% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients.R
\name{predict_0_0}
\alias{predict_0_0}
\title{Build Constraint to Predict f(0) = 0}
\usage{
predict_0_0(it_list, Bp_list, n_bf)
}
\arguments{
\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{n_bf}{Number of basis functions (including the intercept term).}
}
\value{
A \code{vector} to enforce f(0) = 0.
}
\description{
This function constructs the constraint to enforce f(0) = 0.
}
