% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients.R
\name{predict_origin}
\alias{predict_origin}
\title{Build Constraint to Predict f(0) = 0}
\usage{
predict_origin(it_list, Bp_list, n_bf, N)
}
\arguments{
\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{n_bf}{An \code{integer} specifying the number of basis functions, including the intercept term.}

\item{N}{Sample size.}
}
\value{
A \code{vector} of coefficients that enforces the constraint f(0) = 0.
}
\description{
This function constructs the constraint to ensure that the function prediction at zero equals zero, i.e., f(0) = 0.
}
