% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R, R/display_errors.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Fitting}
\usage{
prepare_data(data, x, y, z, degree, error_type)

prepare_data(data, x, y, z, degree, error_type)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{z}{Column indexes of netput variables in \code{data}. These variables are not considered for interaction with other variables.}

\item{degree}{Maximum degree of interaction between variables. It can be a \code{list} of input indexes for interactions or a \code{numeric} value determining the maximum degree of interaction.}

\item{error_type}{A \code{character} string specifying the error structure when fitting the model.}
}
\value{
A \code{matrix} in a [x, z, y] format with variable interactions and / or transformations included.

A \code{matrix} in a [x, z, y] format with variable interactions and / or transformations included.
}
\description{
This function prepares the data for model fitting by generating additional input variables through interactions between variables. It also performs any necessary transformations, such as changing to a logarithmic scale if the error type is multiplicative. It returns a matrix in [x, z, y] format, where x represents input variables, z represents netput variables, and y represents output variables.

This function prepares the data for model fitting by generating additional input variables through interactions between variables. It also performs any necessary transformations, such as changing to a logarithmic scale if the error type is multiplicative. It returns a matrix in [x, z, y] format, where x represents input variables, z represents netput variables, and y represents output variables.
}
