% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackwardAlgorithm.R
\name{PruningAAFS}
\alias{PruningAAFS}
\title{Backward Algorithm for Additive Adaptive Frontier Splines.}
\usage{
PruningAAFS(data, y, AAFS.Forward, d)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{y}{Column output index in \code{data}.}

\item{AAFS.Forward}{\code{list} containing the Forward AAFS model.}

\item{d}{Integer. Generalized Cross Validation (GCV) penalty per knot.}
}
\value{
\code{list} containing a portfolio of Additive Adaptive Frontier Splines sub-models.
}
\description{
This function creates a portfolio of sub-models by removing basis functions one to one.
}
