% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackwardAlgorithm.R
\name{PruningMAFS}
\alias{PruningMAFS}
\title{Backward Algorithm for Multivariate Adaptive Frontier Splines.}
\usage{
PruningMAFS(data, y, ForwardModel, d)
}
\arguments{
\item{data}{\code{Data.frame} or \code{matrix} containing the variables in the model.}

\item{y}{Column output index in \code{data}.}

\item{ForwardModel}{\code{list} containing the Forward MAFS model.}

\item{d}{Integer. Generalized Cross Validation (GCV) penalty per knot. If set to \code{-1}, \code{GCV = RSS / n}.}
}
\value{
\code{list} containing a portfolio of MAFS sub-models.
}
\description{
This function creates a portfolio of sub-models by removing basis functions one to one.
}
