% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{quintic_aces}
\alias{quintic_aces}
\title{Fit a Smoothing Adaptive Constrained Enveloping Splines using Quintic Functions}
\usage{
quintic_aces(
  data,
  x,
  y,
  dea_eff,
  model_type,
  metric,
  monotonicity,
  concavity,
  origin,
  kn_grid,
  kn_side_loc,
  d,
  wq
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{dea_eff}{Indicator vector with 1s for efficient DMUs and 0s for inefficient DMUs.}

\item{model_type}{A \code{character} string specifying the nature of the production frontier that the function will estimate.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{A \code{logical} value indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} value indicating whether to enforce the constraint of concavity in the estimator.}

\item{origin}{A \code{logical} value indicating whether the estimator should satisfy f(0) = 0.}

\item{kn_grid}{A \code{data.frame} containing the set of knots from the backward step for smoothing.}

\item{kn_side_loc}{A \code{list} with the side knot locations for each input dimension.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{wq}{A \code{numeric} value used for the quintic smoothing procedure \insertCite{chen1999}{aces}. This parameter adjusts the distance `d` between the central knot and the left side knot based on the distance `e` between the central knot and the right side knot. If the condition `8/7 < e / d <  1.5` is nos satisfied, then `d` is adjusted to be `wq * e`. This parameter must be set between 8/7 and 1.5.}
}
\value{
A \code{list} containing information relative to the Smoothing Adaptive Constrained Enveloping Splines through quintic functions.
}
\description{
This function performs smoothing on the Adaptive Constrained Enveloping Splines predictor while enforcing continuous second derivatives.
}
\references{
\insertRef{chen1999}{aces}
}
