% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{quintinc_aces}
\alias{quintinc_aces}
\title{Fit a Smoothing Adaptive Constrained Enveloping Splines using Quintic Functions}
\usage{
quintinc_aces(
  data,
  y,
  nX,
  metric,
  monotonicity,
  concavity,
  x0_y0,
  knots,
  x_space,
  wq,
  d
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{y}{Column output indexes in \code{data}.}

\item{nX}{Number of inputs in \code{data}.}

\item{metric}{The lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{knots}{A \code{data.frame} containing the set of knots from backward step for smoothing.}

\item{x_space}{A \code{list} with the central side knot locations in each dimension.}

\item{wq}{For the quintic smoothing procedure \insertCite{chen1999}{aces}. Let `d` be the distance between the central knot and the left side knot, and let be `e` be the distance between the central knot and the right side knot. If the condition `8/7 < e / d < 1.5` is not satisfied, then `d = wq * e`. It must be set between 8/7 and 1.5. If a \code{vector} is entered, the \code{wc} value that most reduce the lack-of-fit criterion is selected.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}
}
\value{
List containing the set of knots from backward (\code{knots}), the new quintic knots (\code{quintic_knots}) and the set of coefficients (\code{coefs}).
}
\description{
This function performs smoothing on the Adaptive Constrained Enveloping Splines predictor while enforcing continuous second derivatives.
}
\references{
\insertRef{chen1999}{aces}
}
