% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{reffcy}
\alias{reffcy}
\title{Random Sample for Efficiency Analysis}
\usage{
reffcy(DGP, parms)
}
\arguments{
\item{DGP}{Data Generation Process:
\itemize {
   \item{\code{"cobb_douglas_XnY1"}}: Cobb-Douglas Data Generation Process for a XnY1 scenario.
   See \code{help("cobb_douglas_XnY1")}.
   \item{\code{"cobb_douglas_X1Y1_noise"}}: Cobb-Douglas Data Generation Process for a X1Y1 scenario with noise.
   See \code{help("cobb_douglas_X1Y1")}.
   \item{\code{"translog_X2Y2"}}: Translog Data Generation Process for a X2Y2 scenario.
   See \code{help("translog_X2Y2")}.
   \item{\code{"add_scenario_XnY1"}}: Additive Data Generation Process for a XnY1 scenario.
   See \code{help("add_scenario_XnY1")}.
   \item{\code{"mult_scenario_XnY1"}}: Multiplicative Data Generation Process for a XnY1 scenario.
   See \code{help("mult_scenario_XnY1")}.
   \item{\code{cobb_douglas_X3Y3}}: Cobb-Douglas Data Generation Process for a X3Y3 scenario.
   See \code{help("cobb_douglas_X3Y3")}.
 }}

\item{parms}{A \code{list} containing the parameters for the simulation functions.
\itemize{
   \item{\code{"N"}}: An integer representing the sample size.
   \item{\code{"nX"}}: An integer representing the number of inputs.
   Applicable  for the \code{cobb_douglas_XnY1(N, nX)} function.
   \item{\code{"border"}}: Proportion of DMUs in the frontier.
   Applicable  for the \code{translog_X2Y2(N, border, noise)} and the \code{cobb_douglas_X3Y3(N, border, noise, returns)} functions.
   \item{\code{"noise"}}: \code{logical} indicating presence of random noise.
   Applicable  for the \code{translog_X2Y2(N, border, noise)} and the \code{cobb_douglas_X3Y3(N, border, noise, returns)} functions.
   \item{\code{"scenario"}}: Determine the type of Data Generation Process.
   Applicable  for the \code{add_scenario_XnY1(N, scenario)} and the \code{mult_scenario_XnY1(N, scenario)} functions.
   \item{\code{"returns"}}: Returns to scale.
   Applicable  for the \code{cobb_douglas_X3Y3(N, border, noise, returns)} function.
   \item{\code{"p"}}: Signal-to-noise ratio.
   Applicable  for the \code{cobb_douglas_X1Y1_noise(N, p)} function.
 }}
}
\value{
A \code{data.frame} simulated with the selected Data Generation Process.
}
\description{
This function simulates a \code{data.frame} with some given Data Generation Process.
}
\details{
Please refer to the help manuals of the mentioned functions for usage examples and more detailed information on the parameters.
}
