% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_scores.R
\name{rf_aces_scores}
\alias{rf_aces_scores}
\title{Compute Efficiency Scores using a Random Forest Adaptive Constrained Enveloping Splines (RF-ACES) model.}
\usage{
rf_aces_scores(eval_data, x, y, object, method, digits = 3)
}
\arguments{
\item{eval_data}{A \code{data.frame} or a \code{matrix} containing the DMUs to be evaluated.}

\item{x}{Column indexes of input variables in \code{tech_data} and \code{eval_data}.}

\item{y}{Column indexes of output and netput variables in \code{tech_data} and \code{eval_data}.}

\item{object}{An \code{rf_aces} object.}

\item{method}{Model prediction used to compute predictions of inputs and obtain a new vector of outputs for \code{eval_data}:
\itemize{
\item{\code{"aces"}}: Random Forest Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_cubic"}}: Random Forest Cubic Smoothed Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_quintic"}}: Random Forest Quintic Smoothed Adaptive Constrained Enveloping Splines model.
}}

\item{digits}{Number of decimal places to which efficiency scores are rounded.}
}
\value{
A \code{data.frame} with the efficiency scores computed through a Random Forest Adaptive Constrained Enveloping Splines (RF-ACES) model.
}
\description{
This function calculates the efficiency scores for each Decision-Making-Unit through a Random Forest Adaptive Constrained Enveloping Splines (RF-ACES) model.
}
