% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{Scenario.C}
\alias{Scenario.C}
\title{Problematic Additive Scenarios}
\usage{
Scenario.C(N, slope)
}
\arguments{
\item{N}{Sample size.}

\item{scenario}{\code{"1"}, \code{"2"}, \code{"3"}, \code{"4"}, \code{"5"}, \code{"6"}, \code{"7"}, \code{"8"}, \code{"9"}, \code{"10"}}
}
\value{
\code{data.frame} with the simulated data.

\insertRef{kuosmanen2010}{mafs}
}
\description{
This function is used to simulate alternative C scenarios.
}
