% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{set_data}
\alias{set_data}
\title{Prepare Data for ACES}
\usage{
set_data(data, x, y, max_degree)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the model variables.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{max_degree}{Maximum interaction degree, or a list of input-index vectors defining the
interactions to add.}
}
\value{
A matrix containing the inputs, input interactions, and outputs, in that order.
}
\description{
Selects the input and output columns and creates the requested multiplicative
input interactions. Interaction names use \code{:}, for example
\code{capital:labour}. The returned matrix places original inputs first,
followed by interaction terms and outputs.
}
