% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{set_interactions}
\alias{set_interactions}
\title{Create Additional Inputs through Interaction Between Variables.}
\usage{
set_interactions(data, x, y, z, degree)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{z}{Column indexes of netput variables in \code{data} (outputs evaluated as inputs).}

\item{degree}{A \code{list} with the input indexes for interaction of variables or a \code{numeric} value that determines the maximum degree of interaction between variables. Basis functions products are constrained to contain factors involving distinct variables to ensure interpretability and avoid multicollinearity.}
}
\value{
A \code{matrix} in a [x,y] format with the interaction of variables included.
}
\description{
This function creates additional inputs through the interaction of the original variables and returns a matrix with the data in a [x,y] format.
}
