% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{set_intervals}
\alias{set_intervals}
\title{Build Input Intervals}
\usage{
set_intervals(Bp_list)
}
\arguments{
\item{Bp_list}{Basis functions grouped by input.}
}
\value{
A list of interval boundaries and active basis function IDs, grouped by input.
}
\description{
Finds the one-dimensional intervals and the basis functions active in each
interval.
}
