% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_algorithm.R
\name{set_knots}
\alias{set_knots}
\title{Find Eligible Knots}
\usage{
set_knots(data, nX, var_idx, minspan, endspan, kn_list, bf, kn_grid)
}
\arguments{
\item{data}{A matrix containing the prepared model variables.}

\item{nX}{Number of inputs.}

\item{var_idx}{Index of the input used to create the new basis functions.}

\item{minspan}{Minimum number of observations between two adjacent knots.}

\item{endspan}{Minimum number of observations before the first knot and after the last knot.}

\item{kn_list}{Selected knots, grouped by input.}

\item{bf}{Parent basis function to expand.}

\item{kn_grid}{Available knot candidates, grouped by input.}
}
\value{
A numeric vector of eligible knots.
}
\description{
Finds knot candidates that satisfy the spacing rules during the forward step.
}
