% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{set_mdim_intervals}
\alias{set_mdim_intervals}
\title{Generate the Set of Multidimensional Intervals}
\usage{
set_mdim_intervals(Bp_list)
}
\arguments{
\item{Bp_list}{A \code{list} containing the set of basis functions by input.}
}
\value{
A \code{list} with the boundaries of the multidimensional intervals.
}
\description{
This function determines the boundaries for the multidimensional intervals.
}
