% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardAlgorithm.R
\name{setKnots}
\alias{setKnots}
\title{Create the set of eligible knots}
\usage{
setKnots(data, nX, xi, L, Le, knots.list, bf, knotsGrid)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{nX}{\code{integer}. Number of inputs.}

\item{xi}{\code{integer}. Index of the variable that creates the new pair of basis functions.}

\item{L}{Minimum number of observations between two adjacent knots.}

\item{Le}{Minimum number of observations before the first and after the final knot.}

\item{knots.list}{\code{list} containing the set of selected knots.}

\item{bf}{\code{list}. Basis function for the expansion of the model.}

\item{knotsGrid}{Grid of virtual knots to perform AAFS.}
}
\value{
Numeric vector with the knots values.
}
\description{
This function generates a vector of knots to create a new pair of basis functions.
}
