% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{side_knot_location}
\alias{side_knot_location}
\title{Find Side-Knot Locations}
\usage{
side_knot_location(data, nX, knots)
}
\arguments{
\item{data}{A matrix containing the prepared model variables.}

\item{nX}{Number of inputs in \code{data}.}

\item{knots}{Data frame of knots retained after pruning.}
}
\value{
A list of central and neighboring knot locations for each input.
}
\description{
Finds the neighboring data values available for side knots around each central
knot.
}
