% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth.R
\name{SmoothForwardMAFS}
\alias{SmoothForwardMAFS}
\title{Smoothing (Forward) Multivariate Adaptive Frontier Splines}
\usage{
SmoothForwardMAFS(data, nX, knots, y)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{nX}{number of inputs in \code{data}.}

\item{knots}{\code{data.frame} containing knots from Forward MAFS.}

\item{y}{output indexes in \code{data}.}
}
\value{
List containing the set of knots from backward (\code{knots}), the new cubic knots (\code{cubic_knots}) and the set of coefficients (\code{alpha}).
}
\description{
This function smoothes the Forward MAFS predictor.
}
