% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth.R
\name{SmoothMAFS}
\alias{SmoothMAFS}
\title{Smoothing Multivariate Adaptive Frontier Splines}
\usage{
SmoothMAFS(data, nX, knots, y)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{nX}{number of inputs in \code{data}.}

\item{knots}{\code{data.frame} containing knots from Backward MAFS.}

\item{y}{output indexes in \code{data}.}
}
\value{
List containing the set of knots from backward (\code{knots}), the new cubic knots (\code{cubic_knots}) and the set of coefficients (\code{alpha}).
}
\description{
This function smoothes the MAFS predictor.
}
