% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{update_Bp_list}
\alias{update_Bp_list}
\title{Rebuild the Basis Function List}
\usage{
update_Bp_list(basis, nX)
}
\arguments{
\item{basis}{Retained basis functions.}

\item{nX}{Number of inputs.}
}
\value{
A list of basis functions grouped by input and side.
}
\description{
Groups the retained basis functions by input and side.
}
