% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{update_knot_list}
\alias{update_knot_list}
\title{Rebuild the Knot List}
\usage{
update_knot_list(Bp_list)
}
\arguments{
\item{Bp_list}{Basis functions grouped by input.}
}
\value{
A list of retained knots.
}
\description{
Rebuilds the knot list from the basis functions retained during pruning.
}
