% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{X3Y3.sim}
\alias{X3Y3.sim}
\title{3 Inputs & 3 Outputs Data Generation}
\usage{
X3Y3.sim(N, noise, inefficiency, return_scale)
}
\arguments{
\item{N}{Sample size.}

\item{noise}{Random noise. 4 possible levels: \code{0}, \code{0.5}, \code{1} or \code{2}.}

\item{inefficiency}{Percentage of inneficient DMUs: \code{0.9}, \code{0.8} or \code{0.7}.}

\item{return_scale}{Returns to scale. \code{"CRS"} for Constant Return to Scale, \code{"DRS"} for Decreasing Return to Scale and \code{"IRS"} for Increasing Return to Scale.}
}
\value{
\code{data.frame} with the simulated data.
}
\description{
This function is used to simulate the data in a scenario with 3 inputs and 3 outputs.
}
