% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{aces}
\alias{aces}
\title{Fit Adaptive Constrained Enveloping Splines (ACES) model}
\usage{
aces(
  data,
  x,
  y,
  y_type = "individual",
  addi_x = NULL,
  degree = 1,
  metric = "mse",
  turbo = Inf,
  monotonicity = TRUE,
  concavity = TRUE,
  x0_y0 = FALSE,
  nterms = 50,
  err_red = 0.01,
  hd_cost = 0.5,
  minspan = 1,
  endspan = 1,
  knots_grid = -1,
  d = 2,
  wc = seq(1, 2, length.out = 5),
  wq = seq(8/7, 1.5, length.out = 5)
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{y_type}{Determines the prediction approach for \code{y}:
\itemize{
\item{\code{"individual"}} to predict one output at a time and use the rest of outputs as additional inputs (netputs).
\item{\code{"all"}} to predict all the outputs at the same time with the original set of inputs.
}}

\item{addi_x}{Column indexes of additional covariables in \code{data}.}

\item{degree}{A \code{list} with the input indexes for interaction of variables or a \code{numeric} value that determines the maximum degree of interaction between variables. Basis functions products are constrained to contain factors involving distinct variables to ensure interpretability and avoid multicollinearity.}

\item{metric}{The lack-of-fit criterion to evaluate the model performance:
\itemize{
\item{\code{"mae"}}: Mean Absolute Error.
\item{\code{"mape"}}: Mean Absolute Percentage Error.
\item{\code{"mse"}}: Mean Squared Error.
\item{\code{"rmse"}}: Root Mean Squared Error.
\item{\code{"nrmse1"}}: Normalized Root Mean Squared Error (using mean).
\item{\code{"nrmse2"}}: Normalized Root Mean Squared Error (using range).
\item{\code{"wape"}}: Weighted Absolute Percentage Error.
}}

\item{turbo}{Hyperparameter aimed at selecting the DMU near the DEA-VRS frontier.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{nterms}{Maximum number of terms created before pruning. Default is \code{50}.}

\item{err_red}{Minimum reduced error rate for the addition of a new pair of 1-degree basis functions. Default is \code{0.01}.}

\item{hd_cost}{Minimum percentage of improvement over the best 1 degree basis function to incorporate a higher degree basis function.}

\item{minspan}{Minimum number of observations between two adjacent knots.
\itemize {
\item{\code{minspan = -2}}: computed as in \insertCite{zhang1994;textual}{aces}.
\item{\code{minspan = -1}}: computed as in \insertCite{friedman1991;textual}{aces}.
\item{\code{minspan = +m}}
}}

\item{endspan}{Minimum number of observations before the first and after the final knot.
\itemize {
\item{\code{endspan = -2}} Computed as in \insertCite{zhang1994;textual}{aces}.
\item{\code{endspan = -1}} Computed as in \insertCite{friedman1991;textual}{aces}.
\item{\code{endspan = +m}}
}}

\item{knots_grid}{Design of the grid of knots. If \code{knotsGrid == -1} (default), the original approach of \insertCite{friedman1991;textual}{aces} based on the observed data is used. Alternatively, a \code{list} with the grid of knots to perform ACES can be entered. Each element of the \code{list} contains a vector with the knots of one variable (e.g., the second element of the list contains the knot values of the second variable and so on).}

\item{d}{Generalized Cross Validation (GCV) penalty per knot. Default is \code{2}.}

\item{wc}{For the cubic smoothing procedure \insertCite{friedman1991}{aces}. Let `d` be the distance between the central knot and the left side knot, and let be `e` be the distance between the central knot and the right side knot. If the condition `1 < d / e < 2` is not satisfied, then `e = wc * d`. It must be set between 1 and 2. If a \code{vector} is entered, the \code{wc} value that most reduce the lack-of-fit criterion is selected.}

\item{wq}{For the quintic smoothing procedure \insertCite{chen1999}{aces}. Let `d` be the distance between the central knot and the left side knot, and let be `e` be the distance between the central knot and the right side knot. If the condition `8/7 < e / d < 1.5` is not satisfied, then `d = wq * e`. It must be set between 8/7 and 1.5. If a \code{vector} is entered, the \code{wc} value that most reduce the lack-of-fit criterion is selected.}
}
\value{
An \code{aces} object.
}
\description{
This function estimates a production frontier satisfying some classical production theory axioms, such as monotonicity and concavity, which is based upon the adaptation of the machine learning technique known as Multivariate Adaptive Regression Splines (MARS) developed by \insertCite{friedman1991;textual}{aces}.
}
\details{
Arguments for the forward algorithm:
\itemize{
\item{\code{nterms}}
\item{\code{err_red}}
\item{\code{minspan}}
\item{\code{endspan}}
\item{\code{knotsGrid}}
}

Arguments for the backward algorithm:
\itemize{
\item{\code{d}}
}

Arguments for the smoothing procedure:
\itemize{
\item{\code{wc}}
\item{\code{wq}}
}
}
\references{
\insertRef{zhang1994}{aces} \cr
\cr
\insertRef{friedman1991}{aces} \cr
\cr
\insertRef{chen1999}{aces}
}
