% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{aces_algorithm}
\alias{aces_algorithm}
\title{Algorithm of Adaptive Constrained Enveloping Splines (ACES)}
\usage{
aces_algorithm(
  data,
  x,
  y,
  z,
  y_type,
  addi_x,
  degree,
  metric,
  monotonicity,
  concavity,
  x0_y0,
  nterms,
  err_red,
  hd_cost,
  minspan,
  endspan,
  knots_grid,
  d,
  wc,
  wq
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{z}{Column indexes of netput variables in \code{data} (outputs evaluated as inputs). These variables must be considered as output when computing predictions or efficiency scores.}

\item{y_type}{Output prediction strategy: whether it should be done all at once or one at a time.}

\item{addi_x}{Column indexes of additional covariables in \code{data}.}

\item{degree}{Maximum degree of interaction between variables.}

\item{metric}{The lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{nterms}{For the Forward algorithm. Maximum number of terms created before pruning.}

\item{err_red}{For the Forward algorithm. Minimum reduced error rate for the addition of a new pair of 1-degree basis functions.}

\item{hd_cost}{Minimum percentage of improvement over the best 1 degree basis function to incorporate a higher degree basis function.}

\item{minspan}{For the Forward algorithm. Minimum number of observations between two adjacent knots.}

\item{endspan}{For the Forward algorithm. Minimum number of observations before the first and after the final knot.}

\item{knots_grid}{For the Forward algorithm. Design of the grid of knots.}

\item{d}{For the Backward algorithm. Generalized Cross Validation (GCV) penalty per knot.}

\item{wc}{Hyperparameter for the side knot distances in the cubic smoothing procedure \insertCite{friedman1991}{aces}.}

\item{wq}{Hyperparameter for the side knot distances in the quintic smoothing procedure \insertCite{chen1999}{aces}.}
}
\value{
An \code{aces} object.
}
\description{
This function estimates a production frontier satisfying some classical production theory axioms, such as monotonicity and concavity, which is based upon the adaptation of the machine learning technique known as Multivariate Adaptive Regression Splines (MARS) developed by \insertCite{friedman1991;textual}{aces}.
}
\references{
\insertRef{zhang1994}{aces} \cr
\cr
\insertRef{friedman1991}{aces} \cr
\cr
\insertRef{chen1999}{aces}
}
