% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{aces_algorithm}
\alias{aces_algorithm}
\title{Algorithm of Adaptive Constrained Enveloping Splines (ACES).}
\usage{
aces_algorithm(
  data,
  x_vars,
  y_vars,
  z_vars,
  quick_aces,
  error_type,
  max_degree,
  compl_cost,
  metric,
  shape,
  nterms,
  err_red,
  minspan,
  endspan,
  kn_grid,
  kn_penalty
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x_vars}{Column indexes of input variables in \code{data}.}

\item{y_vars}{Column indexes of output variables in \code{data}.}

\item{z_vars}{Column indexes of contextual variables in \code{data}.}

\item{quick_aces}{A \code{logical} indicating if the fast version of ACES should be employed.}

\item{error_type}{A \code{character} string specifying the error structure when fitting the model.}

\item{max_degree}{Maximum degree of interaction between variables.}

\item{compl_cost}{Minimum percentage of improvement over the best 1 degree BF to incorporate a higher degree BF.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion employed to evaluate the model performance.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity and/or passing through the origin (only for piece-wise linear version).}

\item{nterms}{Maximum number of terms created during the forward step.}

\item{err_red}{Minimum reduced error rate for the addition of a new pair of 1-degree BFs.}

\item{minspan}{Minimum number of observations between two adjacent knots.}

\item{endspan}{Minimum number of observations before the first and after the final knot.}

\item{kn_grid}{Grid design for knots placement in ACES.}

\item{kn_penalty}{Penalty per knot for computing Generalized Cross Validation.}
}
\value{
An \code{aces} object.
}
\description{
This function implements the Adaptive Constrained Enveloping Splines (ACES) algorithm.
}
