% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{aces_object}
\alias{aces_object}
\title{Create an aces object}
\usage{
aces_object(
  data,
  x,
  y,
  z,
  y_type,
  addi_x,
  degree,
  metric,
  monotonicity,
  concavity,
  x0_y0,
  nterms,
  err_red,
  minspan,
  endspan,
  kn_grid,
  d,
  wc,
  wq,
  aces_forward,
  aces,
  aces_cubic,
  aces_quintic
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{z}{Column indexes of netput variables in \code{data} (outputs evaluated as inputs).}

\item{y_type}{Determines the prediction approach for \code{y}: whether it should be done all at once or one at a time.}

\item{addi_x}{Column indexes of additional covariables in \code{data}.}

\item{degree}{Maximum degree of interaction between variables.}

\item{metric}{The lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{nterms}{Maximum number of terms created by the forward algorithm (before pruning).}

\item{err_red}{Minimum reduced error rate for the addition of two new basis functions.}

\item{minspan}{Minimum number of observations between two adjacent knots.}

\item{endspan}{Minimum number of observations before the first and after the final knot.}

\item{kn_grid}{Grid of knots to perform ACES.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{wc}{Hyperparameter for the side knot distances in the cubic smoothing procedure \insertCite{friedman1991}{aces}.}

\item{wq}{Hyperparameter for the side knot distances in the quintic smoothing procedure \insertCite{chen1999}{aces}.}

\item{aces_forward}{A \code{list} containing the forward Adaptive Constrained Enveloping Splines model.}

\item{aces}{A \code{list} containing the Adaptive Constrained Enveloping Splines model.}

\item{aces_cubic}{A \code{list} containing the smoothed version of Adaptive Constrained Enveloping Splines through cubic basis functions.}

\item{aces_quintic}{A \code{list} containing the smoothed version of Adaptive Constrained Enveloping Splines through quintic basis functions.}
}
\value{
An \code{aces} object.
}
\description{
This function saves information about the Adaptive Constrained Enveloping Splines (ACES) model.
}
\references{
\insertRef{friedman1991}{aces} \cr
\cr
\insertRef{chen1999}{aces}
}
