% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{aces_object}
\alias{aces_object}
\title{Create an aces object}
\usage{
aces_object(
  data,
  x,
  y,
  z,
  error_type,
  max_degree,
  compl_cost,
  xi_degree,
  metric,
  shape,
  nterms,
  err_red,
  minspan,
  endspan,
  kn_grid,
  kn_penalty,
  wc,
  wq,
  aces_forward,
  aces,
  aces_cubic,
  aces_quintic,
  technology
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{z}{Column indexes of contextual variables in \code{data}.}

\item{error_type}{A \code{character} string specifying the error structure when fitting the model.}

\item{max_degree}{Maximum degree of interaction between variables.}

\item{compl_cost}{Minimum percentage of improvement over the best 1 degree basis function to incorporate a higher degree basis function.}

\item{xi_degree}{A \code{matrix} specifying the degree of each variable.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity.}

\item{nterms}{Maximum number of terms created before pruning.}

\item{err_red}{Minimum reduced error rate for the addition of a new pair of 1-degree basis functions.}

\item{minspan}{Minimum number of observations between two adjacent knots.}

\item{endspan}{Minimum number of observations before the first and after the final knot.}

\item{kn_grid}{Design of the grid of knots to perform ACES}

\item{kn_penalty}{Generalized Cross Validation (GCV) penalty per knot.}

\item{wc}{Hyperparameter for the side knot distances in the cubic smoothing procedure.}

\item{wq}{Hyperparameter for the side knot distances in the quintic smoothing procedure.}

\item{aces_forward}{A \code{list} containing the forward step of the Adaptive Constrained Enveloping Splines model.}

\item{aces}{A \code{list} containing the Adaptive Constrained Enveloping Splines model.}

\item{aces_cubic}{A \code{list} containing the Smooth Adaptive Constrained Enveloping Splines through cubic basis functions.}

\item{aces_quintic}{A \code{list} containing the Smooth Adaptive Constrained Enveloping Splines through quintic basis functions.}

\item{technology}{A \code{list} with the points that make up the technology set for each model.}
}
\value{
An \code{aces} object.
}
\description{
This function saves information about the Adaptive Constrained Enveloping Splines (ACES) model.
}
