% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{aces_pruning}
\alias{aces_pruning}
\title{Backward Algorithm for Adaptive Constrained Enveloping Splines}
\usage{
aces_pruning(
  data,
  y,
  xi_degree,
  dea_eff,
  model_type,
  metric,
  forward_model,
  Bp_list,
  monotonicity,
  concavity,
  origin,
  d
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{xi_degree}{A \code{matrix} indicating the degree of each input variable.}

\item{dea_eff}{An indicator vector with 1s for efficient DMUs and 0s for inefficient DMUs.}

\item{model_type}{A \code{character} string specifying the nature of the production frontier that the function will estimate.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{forward_model}{A \code{list} containing the information relative to the forward step of the ACES model.}

\item{Bp_list}{A \code{list} containing the current set of basis functions for each input variable.}

\item{monotonicity}{A \code{logical} value indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} value indicating whether to enforce the constraint of concavity in the estimator.}

\item{origin}{A \code{list} containing the current set of selected knots for each input variable.}

\item{d}{An \code{integer} specifying the Generalized Cross Validation (GCV) penalty per knot.}
}
\value{
A \code{list} containing a portfolio of Adaptive Constrained Enveloping Splines sub-models.
}
\description{
This function implements the Backward Algorithm in Adaptive Constrained Enveloping Splines. It creates a portfolio of sub-models by iteratively removing basis functions one by one.
}
