% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{aces_pruning}
\alias{aces_pruning}
\title{Backward Algorithm for Adaptive Constrained Enveloping Splines}
\usage{
aces_pruning(
  data,
  xi_degree,
  y,
  metric,
  monotonicity,
  concavity,
  x0_y0,
  forward_model,
  Bp_list,
  d
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{xi_degree}{Matrix indicating the degree of each input variable.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{metric}{Lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{forward_model}{\code{list} containing the Forward ACES model.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{d}{An \code{integer} specifying the Generalized Cross Validation (GCV) penalty per knot.}
}
\value{
A \code{list} containing a portfolio of Adaptive Constrained Enveloping Splines sub-models.
}
\description{
This function implements the Backward Algorithm in Adaptive Constrained Enveloping Splines to create a portfolio of sub-models by iteratively removing basis functions one by one.
}
