% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{aces_pruning}
\alias{aces_pruning}
\title{Backward Algorithm for Adaptive Constrained Enveloping Splines}
\usage{
aces_pruning(
  data,
  x,
  y,
  xi_degree,
  model_type,
  dea_scores,
  metric,
  forward_model,
  Bp_list,
  shape,
  kn_penalty
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{xi_degree}{A \code{matrix} indicating the degree of each input variable.}

\item{model_type}{A \code{character} string specifying the nature of the production frontier.}

\item{dea_scores}{A \code{matrix} containing DEA-VRS efficiency scores, calculated using an output-oriented radial model. For models with multiple outputs, each column corresponds to the scores for one specific output.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{forward_model}{A \code{list} containing the information relative to the forward step of the ACES model.}

\item{Bp_list}{A \code{list} containing the current set of basis functions for each input variable.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity and/or passing through the origin.}

\item{kn_penalty}{An \code{integer} specifying the Generalized Cross Validation (GCV) penalty per knot.}
}
\value{
A \code{list} containing a portfolio of Adaptive Constrained Enveloping Splines sub-models.
}
\description{
This function implements the Backward Algorithm in Adaptive Constrained Enveloping Splines. It creates a portfolio of sub-models by iteratively removing basis functions one by one.
}
