% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_algorithm.R
\name{add_basis_function}
\alias{add_basis_function}
\title{Add a New Pair of Basis Functions in Adaptive Constrained Enveloping Splines}
\usage{
add_basis_function(
  data,
  x,
  xi_degree,
  y,
  metric,
  monotonicity,
  concavity,
  x0_y0,
  forward_model,
  kn_list,
  Bp_list,
  L,
  Le,
  kn_grid,
  err_min,
  hd_cost
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{xi_degree}{Matrix indicating the degree of each input variable.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{metric}{Lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{forward_model}{A \code{list} containing the current set of basis functions (\code{BF_set}) and the B matrix (\code{B}).}

\item{kn_list}{A \code{list} containing the current set of selected knots for each input variable.}

\item{Bp_list}{A \code{list} containing the current set of basis functions for each input variable.}

\item{L}{Minimum number of observations between two adjacent knots.}

\item{Le}{Minimum number of observations before the first and after the final knot.}

\item{kn_grid}{Grid of knots to perform ACES.}

\item{err_min}{Minimum error obtained by the forward algorithm in the current iteration.}

\item{hd_cost}{Minimum percentage of improvement over the best 1 degree basis function to incorporate a higher degree basis function.}
}
\value{
An undated \code{list} containing the matrix of basis functions (\code{B}), an undated \code{list} with information of the basis functions (\code{BF_set}), an updated \code{list} of selected knots (\code{knots_list}) and the minimum error obtained (\code{err_min}).
}
\description{
This function adds the pair of basis functions that results in the largest reduction of the lack-of-fit criterion.
}
