% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_algorithm.R
\name{add_basis_function}
\alias{add_basis_function}
\title{Add a New Pair of Basis Functions in Adaptive Constrained Enveloping Splines}
\usage{
add_basis_function(
  data,
  x,
  y,
  xi_degree,
  dea_eff,
  model_type,
  metric,
  forward_model,
  Bp_list,
  monotonicity,
  concavity,
  origin,
  kn_list,
  kn_grid,
  L,
  Le,
  err_min,
  hd_cost
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{xi_degree}{A \code{matrix} indicating the degree of each input variable.}

\item{dea_eff}{An indicator vector with 1s for efficient DMUs and 0s for inefficient DMUs.}

\item{model_type}{A \code{character} string specifying the nature of the production frontier that the function will estimate.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{forward_model}{A \code{list} containing the current set of basis functions (\code{BF_set}) and the B matrix (\code{B}).}

\item{Bp_list}{A \code{list} containing the current set of basis functions for each input variable.}

\item{monotonicity}{A \code{logical} value indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} value indicating whether to enforce the constraint of concavity in the estimator.}

\item{origin}{A \code{logical} value indicating whether the estimator should satisfy f(0) = 0.}

\item{kn_list}{A \code{list} containing the current set of selected knots for each input variable.}

\item{kn_grid}{A \code{list} with the grid of knots to perform ACES.}

\item{L}{A \code{integer} value specifying the minimum number of observations between two adjacent knots.}

\item{Le}{A \code{integer} value specifying the minimum number of observations before the first and after the final knot.}

\item{err_min}{A \code{numeric} value specifying the minimum error obtained by the forward algorithm in the current iteration.}

\item{hd_cost}{A \code{numeric} value specifying the minimum percentage of improvement over the best 1 degree basis function to incorporate a higher degree basis function.}
}
\value{
An updated \code{list} containing the matrix of basis functions (\code{B}), an updated \code{list} with information of the basis functions (\code{BF_set}), an updated \code{list} of selected knots (\code{knots_list}) and the minimum error obtained (\code{err_min}).
}
\description{
This function adds the pair of basis functions that results in the largest reduction of the lack-of-fit criterion.
}
