% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForwardAlgorithm.R
\name{AddBF}
\alias{AddBF}
\title{Add a New Pair of Basis Functions}
\usage{
AddBF(
  data,
  x,
  y,
  ForwardModel,
  knots.list,
  cand.knots,
  Kp,
  L,
  Le,
  knotsGrid,
  err.min
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{ForwardModel}{\code{list} containing the set of basis functions (\code{BF}) and the B matrix (\code{B}).}

\item{knots.list}{\code{list} containing the set of selected knots.}

\item{cand.knots}{\code{list} containing the candidate knots.}

\item{Kp}{Maximum degree of interaction allowed.}

\item{L}{Minimum number of observations between two adjacent knots.}

\item{Le}{Minimum number of observations before the first and after the final knot.}

\item{knotsGrid}{Grid of virtual knots to perform AAFS.}

\item{err.min}{Forward algorithm minimum error in the given iteration.}
}
\value{
A \code{list} containing the matrix of basis functions (\code{B}), a \code{list} of basis functions (\code{BF}), a \code{list} of selected knots (\code{knots.list}) and the minimum given error (\code{err.min}).
}
\description{
This function adds the pair of basis functions that produce the largest reduction in the lack-of-fit.
}
