% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{basis_to_drop}
\alias{basis_to_drop}
\title{Basis Functions to be Dropped}
\usage{
basis_to_drop(Bp_list, conc, nX)
}
\arguments{
\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{conc}{A \code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{nX}{Number of inputs.}
}
\value{
A \code{data.frame} indicating the candidate basis function to be dropped for each input.
}
\description{
This function determines the basis function to be dropped from a set of intervals.
}
