% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cet_cd_X2Y2}
\alias{cet_cd_X2Y2}
\title{2 outputs ~ 2 inputs Constant Elasticity of Transformation - Cobb-Douglas Data Generation Process}
\usage{
cet_cd_X2Y2(N, border, noise)
}
\arguments{
\item{N}{An integer representing the sample size.}

\item{border}{Proportion of DMUs in the frontier.}

\item{noise}{A \code{logical} indicating whether to add random noise.}
}
\value{
A \code{data.frame} with the simulated data: 2 inputs (x1, x2), 2 outputs (y1, y2), the theoretical frontier (yT1, yT2).
}
\description{
This function simulates a 2 outputs ~ 2 inputs \code{data.frame} using a Data Generation Process (DGP) based on
Constant Elasticity of Transformation (CET) on the output side and Cobb-Douglas (CD) on the input side as in \insertCite{fare1994;textual}{aces}.

The dataset consists of \code{N} observations divided into four groups with equal proportions:
\itemize{
  \item \strong{Small-size producers}: \eqn{h(u) = 25}, \eqn{f(x) = 25}, \eqn{\delta = 0.898}
  \item \strong{Medium-size producers (Lower bound)}: \eqn{h(u) = 50}, \eqn{f(x) = 50}, \eqn{\delta = 1.000}
  \item \strong{Medium-size producers (Upper bound)}: \eqn{h(u) = 75}, \eqn{f(x) = 75}, \eqn{\delta = 1.000}
  \item \strong{Large-size producers}: \eqn{h(u) = 100}, \eqn{f(x) = 100}, \eqn{\delta = 0.927}
}

This structure ensures the artificial technology exhibits regions of increasing, constant, and decreasing returns to scale.
}
\references{
\insertRef{fare1994}{aces}
}
