% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{change_technology}
\alias{change_technology}
\title{Update Technology Set}
\usage{
change_technology(
  object,
  tech_xmat = NULL,
  tech_ymat = NULL,
  psi = NULL,
  pto = FALSE
)
}
\arguments{
\item{object}{An \code{aces} object.}

\item{tech_xmat}{A \code{matrix} representing the input data (optional). Required if \code{psi} is provided.}

\item{tech_ymat}{A \code{matrix} representing the output data (optional). Required if \code{psi} is provided.}

\item{psi}{A \code{numeric} threshold controlling the refinement of predicted outputs (optional). If the predicted value for a given output exceeds the observed value by more than \code{psi}, the observed value is used instead.}

\item{pto}{A \code{logical} indicating if (0,0) should be added to the technology set. Default is \code{FALSE}.}
}
\value{
An \code{aces} object with updated technologies.
}
\description{
This function constructs a new production technology by modifying the output matrix based on a refinement procedure that replaces overestimated outputs. Additionally, it allows incorporating the origin into the technology set.
}
