% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{change_technology}
\alias{change_technology}
\title{Generate a New Technology Set}
\usage{
change_technology(object, tech_xmat, tech_ymat, psi)
}
\arguments{
\item{object}{An \code{aces} object.}

\item{tech_xmat}{A \code{matrix} representing the input data.}

\item{tech_ymat}{A \code{matrix} representing the output data.}

\item{psi}{A \code{numeric} threshold controlling the refinement of predicted outputs. If the predicted value for a given output exceeds the observed value by more than \code{psi}, the observed value is used instead.}
}
\value{
An \code{aces} object with updated technologies.
}
\description{
This function constructs a new production technology by modifying the output matrix based on a refinement
procedure that replaces overestimated outputs. For each output, the predicted values from an ACES model
(tech_ymat2) are compared against the original observed values (tech_ymat1). When the predicted value exceeds
the observed one by more than a predefined threshold, the observed value is retained instead.
}
