% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cobb_douglas_XnY1_noise}
\alias{cobb_douglas_XnY1_noise}
\title{1 output ~ nX input Cobb-Douglas Data Generation Process with Noise}
\usage{
cobb_douglas_XnY1_noise(N, nX, p, heteroskedasticity = FALSE)
}
\arguments{
\item{N}{An integer representing the sample size.}

\item{nX}{An integer representing the number of inputs.}

\item{p}{Noise-to-signal ratio.}

\item{heteroskedasticity}{A \code{"logical"} value indicating if heteroskedasticity should be introduced in the inefficiency term. Only available for 3 inputs (\code{nX = 3}).}
}
\value{
A \code{data.frame} with the simulated data: 1 inputs (x1), 1 output (y) and the theoretical frontier (yT).
}
\description{
This function simulates a 1 output ~ nX input \code{data.frame} with a Cobb-Douglas Data Generation Process including statistical noise as in \insertCite{simar2011;textual}{aces}.
}
\references{
\insertRef{simar2011}{aces}
}
\keyword{internal}
