% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{compute_gcv}
\alias{compute_gcv}
\title{Compute Generalized Cross-Validation (GCV)}
\usage{
compute_gcv(y_obs, LOF, n_bf, kn_penalty, kn_list, xi_degree)
}
\arguments{
\item{y_obs}{Vector of observed data.}

\item{LOF}{A \code{numeric} value representing the lack-of-fit of the model. This term quantifies how well the model fits the data.}

\item{n_bf}{A \code{integer} specifying the number of basis functions used in the model.}

\item{kn_penalty}{A \code{numeric} value representing the penalty applied for each knot in the model.}

\item{kn_list}{A \code{list} containing the set of knots used in the model.}

\item{xi_degree}{A \code{matrix} indicating the degree of each input variable.}
}
\value{
A \code{numeric} value representing the computed Generalized Cross-Validation (GCV) score.
}
\description{
This function computes the generalized cross-validation for the backward algorithm in Adaptive Constrained Enveloping Splines (ACES).
}
