% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{compute_gcv}
\alias{compute_gcv}
\title{Compute Generalized Cross-Validation (GCV)}
\usage{
compute_gcv(y_obs, y_hat, metric, B, d, knots, xi_degree)
}
\arguments{
\item{y_obs}{Vector of observed data.}

\item{y_hat}{Vector of predicted values.}

\item{metric}{Lack-of-fit criterion to evaluate the model performance.}

\item{B}{Matrix of basis functions.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{knots}{\code{list} with the set of knots.}

\item{xi_degree}{Matrix indicating the degree of each input variable.}
}
\value{
The Generalized Cross-Validation value.
}
\description{
This function computes the generalized cross-validation for the Backward Algorithm in Adaptive Constrained Enveloping Splines.
}
