% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{compute_gcv}
\alias{compute_gcv}
\title{Compute Generalized Cross-Validation (GCV)}
\usage{
compute_gcv(y_obs, y_hat, metric, n_bf, d, knots, xi_degree)
}
\arguments{
\item{y_obs}{A \code{numeric} matrix of observed data.}

\item{y_hat}{A \code{numeric} matrix of predicted values.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{n_bf}{A \code{integer} specifying the number of basis functions in the model.}

\item{d}{A \code{numeric} value specifying the Generalized Cross Validation (GCV) penalty per knot.}

\item{knots}{A \code{list} specifying the set of knots.}

\item{xi_degree}{A \code{matrix} indicating the degree of each input variable.}
}
\value{
The Generalized Cross-Validation (GCV) value.
}
\description{
This function computes the generalized cross-validation for the backward algorithm in Adaptive Constrained Enveloping Splines.
}
