% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{compute_span}
\alias{compute_span}
\title{Compute Minimum and End Spans}
\usage{
compute_span(data, minspan, endspan, nX)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{minspan}{Minimum number of observations between two adjacent knots.
\itemize{
\item{\code{minspan = -2}} Computed as in \insertCite{zhang1994;textual}{aces}.
\item{\code{minspan = -1}} Computed as in \insertCite{friedman1991;textual}{aces}.
\item{\code{minspan = +m}}
}}

\item{endspan}{Minimum number of observations before the first and after the final knot.
\itemize{
\item{\code{endspan = -2}} Computed as in \insertCite{zhang1994;textual}{aces}.
\item{\code{endspan = -1}} Computed as in \insertCite{friedman1991;textual}{aces}.
\item{\code{endspan = +m}}
}}

\item{nX}{Number of inputs.}
}
\value{
A \code{list} with the minimum span and the end span.
}
\description{
This function computes the minimum span (i.e., the minimum number of observations between two adjacent knots) and the end span (i.e., the minimum number of observations before the first and after the final knot).
}
