% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R, R/display_errors.R
\name{compute_span}
\alias{compute_span}
\title{Compute Minimum and End Span}
\usage{
compute_span(data, minspan, endspan, nX)

compute_span(data, minspan, endspan, nX)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{minspan}{A \code{numeric} value or vector specifying the minimum number of observations between two adjacent knots. The following options are available:
\itemize{
  \item{\code{minspan = -2}}: Computed according to the method proposed by \insertCite{zhang1994;textual}{aces}.
  \item{\code{minspan = -1}}: Computed according to the method proposed by \insertCite{friedman1991;textual}{aces}.
  \item{\code{minspan = +m}}: A positive integer specifying the exact number of observations.
}}

\item{endspan}{A \code{numeric} value or vector specifying the minimum number of observations before the first knot and after the final knot. The following options are available:
\itemize{
  \item{\code{endspan = -2}}: Computed according to the method proposed by \insertCite{zhang1994;textual}{aces}.
  \item{\code{endspan = -1}}: Computed according to the method proposed by \insertCite{friedman1991;textual}{aces}.
  \item{\code{endspan = +m}}: A positive integer specifying the exact number of observations.
}}

\item{nX}{Number of input variables.}
}
\value{
A \code{list} with two components:
\itemize{
  \item{\code{minspan}}: The computed minimum span.
  \item{\code{endspan}}: The computed end span.
}

A \code{list} with two components:
\itemize{
  \item{\code{minspan}}: The computed minimum span.
  \item{\code{endspan}}: The computed end span.
}
}
\description{
This function computes the minimum span, which is the minimum number of observations between two adjacent knots, and the end span, which is the minimum number of observations before the first knot and after the final knot.

This function computes the minimum span, which is the minimum number of observations between two adjacent knots, and the end span, which is the minimum number of observations before the first knot and after the final knot.
}
\references{
\insertRef{friedman1991}{aces} \cr \cr
\insertRef{zhang1994}{aces}

\insertRef{friedman1991}{aces} \cr \cr
\insertRef{zhang1994}{aces}
}
