% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAFS.R
\name{computeSpan}
\alias{computeSpan}
\title{Compute Span}
\usage{
computeSpan(data, minspan, endspan, nX)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{minspan}{Minimum number of observations between knots.
\itemize{
\item{\code{minspan = -2}} Minspan computed as in Zhang (1994).
\item{\code{minspan = -1}} Minspan computed as in Friedman's MARS paper section 3.8 with alpha = 0.05.
\item{\code{minspan =  m}} \code{m} observations between knots.
}}

\item{endspan}{Minimum number of observations before the first and after the final knot.
\itemize{
\item{\code{endspan = -2}} Endspan computed as in Zhang (1994).
\item{\code{endspan = -1}} Endspan computed as in Friedman's MARS paper section 3.8 with alpha = 0.05.
\item{\code{endspan =  m}} \code{m} observations between knots.
}}

\item{nX}{Number of inputs.}
}
\value{
\code{list} with the minimum span and the end span.
}
\description{
This function computes the minimum span and the end span.
}
