% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{create_cubic_basis}
\alias{create_cubic_basis}
\title{Generate a new pair of Cubic Basis Functions}
\usage{
create_cubic_basis(data, xi, knots, B, side)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{xi}{Variable index of the variable that creates the new basis function(s).}

\item{knots}{Triplet of knots for creating the new basis function(s).}

\item{B}{A \code{matrix} of basis functions.}

\item{side}{Side of the new basis function.}
}
\value{
A \code{matrix} of basis functions updated with the new cubic basis functions.
}
\description{
This function generates two new cubic basis functions from a variable and a triplet of knots.
}
