% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_algorithm.R
\name{create_linear_basis}
\alias{create_linear_basis}
\title{Generate a Pair of Linear Basis Functions}
\usage{
create_linear_basis(data, var_idx, knot)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{var_idx}{Index of the input variable that creates the new pair of basis functions.}

\item{knot}{Knot for creating the new pair of basis functions.}
}
\value{
A \code{list} with the new pair of basis functions.
}
\description{
This function generates two new linear basis functions from a variable and a knot.
}
