% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{create_quintic_basis}
\alias{create_quintic_basis}
\title{Generate a New Pair of Quintic Basis Functions}
\usage{
create_quintic_basis(data, xi, knots, B, side)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{xi}{Index of the variable that creates the new basis function(s).}

\item{knots}{Triplet of knots for creating the new basis function(s).}

\item{B}{A \code{matrix} of basis functions.}

\item{side}{Side of the new basis function inherited from the linar basis function.}
}
\value{
A \code{matrix} of basis functions updated with the new quintic basis functions.
}
\description{
This function generates two new quintic basis functions from a variable and a triplet of knots.
}
