% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{cubic_aces}
\alias{cubic_aces}
\title{Fit a Smoothing Adaptive Constrained Enveloping Splines using Cubic Functions}
\usage{
cubic_aces(
  data,
  y,
  nX,
  metric,
  monotonicity,
  concavity,
  x0_y0,
  knots,
  x_space,
  wc,
  d
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{y}{Column output indexes in \code{data}.}

\item{nX}{Number of inputs in \code{data}.}

\item{metric}{The lack-of-fit criterion to evaluate the model performance.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}

\item{knots}{A \code{data.frame} containing the set of knots from backward step for smoothing.}

\item{x_space}{A \code{list} with the central side knot locations in each dimension.}

\item{wc}{For the cubic smoothing procedure \insertCite{friedman1991}{aces}. Let `d` be the distance between the central knot and the left side knot, and let be `e` be the distance between the central knot and the right side knot. If the condition `1 < d / e < 2` is not satisfied, then `e = wc * d`. It must be set between 1 and 2. If a \code{vector} is entered, the \code{wc} value that most reduce the lack-of-fit criterion is selected.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}
}
\value{
A \code{list} containing the information of the Smoothing Adaptive Constrained Enveloping Splines through cubic functions.
}
\description{
This function performs smoothing on the Adaptive Constrained Enveloping Splines predictor while enforcing continuous first derivatives.
}
\references{
\insertRef{friedman1991}{aces}
}
