% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{cubic_aces}
\alias{cubic_aces}
\title{Fit a Smooth Adaptive Constrained Enveloping Splines using Cubic Functions}
\usage{
cubic_aces(
  data,
  x,
  y,
  dea_scores,
  model_type,
  metric,
  shape,
  kn_grid,
  kn_side_loc,
  kn_penalty,
  xi_degree,
  wc
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{dea_scores}{An indicator vector with 1s for efficient DMUs and 0s for inefficient DMUs.}

\item{model_type}{A \code{character} string specifying the nature of the production frontier that the function will estimate.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity.}

\item{kn_grid}{A \code{data.frame} containing the set of knots from the backward step for smoothing.}

\item{kn_side_loc}{A \code{list} with the side knot locations for each input dimension.}

\item{kn_penalty}{Generalized Cross Validation (GCV) penalty per knot.}

\item{xi_degree}{A \code{matrix} indicating the degree of each input variable.}

\item{wc}{Let `p` be the distance between the central knot and the right-side knot, and `v` be the distance between the central knot and the left-side knot during the smoothing procedure.A \code{numeric} value used for cubic smoothing \insertCite{friedman1991}{aces}. This parameter is defined as `v / p` and must be set between 1 and 2. If a \code{vector} is entered, the \code{wc} value that most reduced the lack-of-fit criterion is selected.}
}
\value{
A \code{list} containing information relative to the Smooth Adaptive Constrained Enveloping Splines through cubic functions.
}
\description{
This function performs smoothing on the Adaptive Constrained Enveloping Splines predictor while enforcing continuous first derivatives.
}
\references{
\insertRef{friedman1991}{aces}
}
