% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_errors.R
\name{display_errors_aces}
\alias{display_errors_aces}
\title{Error Messaging in Adaptive Constrained Enveloping Splines (ACES) functions}
\usage{
display_errors_aces(
  data,
  x,
  y,
  quick_aces,
  error_type,
  max_degree,
  compl_cost,
  metric,
  nterms,
  err_red,
  minspan,
  endspan,
  kn_grid,
  kn_penalty
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{quick_aces}{A \code{logical} indicating if the fast version of ACES should be employed.}

\item{error_type}{A \code{character} string specifying the error structure when fitting the model. It can be either: \code{"add"} or \code{"mul"}.}

\item{max_degree}{A \code{list} with input indexes for interaction of variables, or a \code{numeric} specifying the maximum max_degree of interaction.}

\item{compl_cost}{A \code{numeric} value specifying the minimum percentage of improvement over the best 1-degree BF to incorporate a higher degree BF.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion employed to evaluate the model performance. Options are: \code{"mae"}, \code{"mape"}, \code{"mse"}, \code{"rmse"}, \code{"nrmse1"} or \code{"nrmse2"}.}

\item{nterms}{A positive \code{integer} specifying the maximum number of terms created during the forward step.}

\item{err_red}{A \code{numeric} value specifying the minimum reduced error rate for the addition of a new pair of 1-degree BFs.}

\item{minspan}{A \code{numeric} value specifying the minimum number of observations between two adjacent knots. Options are: \code{"-2"}, \code{"-1"} or \code{"m"}.}

\item{endspan}{A \code{numeric} value specifying the minimum number of observations before the first and after the final knot. Options are: \code{"-2"}, \code{"-1"} or \code{"m"}.}

\item{kn_grid}{Grid of knots to perform ACES. It can be: \code{-1} or a \code{list}.}

\item{kn_penalty}{A positive \code{numeric} value specifying the Generalized Cross Validation penalty per knot.}

\item{model_type}{A \code{character} string specifying the nature of the estimated production frontier. It can be either: \code{"envelopment"} or \code{"stochastic"}.}
}
\value{
This function returns error messages if hyperparameters are incorrectly specified.
}
\description{
This function displays error messages if hyperparameters are incorrectly specified in ACES functions aimed at estimating production functions.
}
