% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_errors.R
\name{display_errors_rf_aces}
\alias{display_errors_rf_aces}
\title{Error Messaging in Random-Forest Adaptive Constrained Enveloping Splines (RF-ACES) functions}
\usage{
display_errors_rf_aces(
  data,
  x,
  y,
  quick_aces,
  error_type,
  learners,
  nvars,
  sample_size,
  max_degree,
  compl_cost,
  metric,
  nterms,
  err_red,
  minspan,
  endspan,
  kn_grid
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{quick_aces}{A \code{logical} indicating if the fast version of ACES should be employed.}

\item{error_type}{A \code{character} string specifying the error structure that the function will use when fitting the model. It can be either: \code{"add"} or \code{"mul"}.}

\item{learners}{An \code{integer} indicating the number of models for bagging.}

\item{nvars}{An \code{integer} indicating the number of variables randomly chosen at each split in RF-ACES.}

\item{sample_size}{An \code{integer} indicating the number of samples to draw from \code{data} to train each base estimator.}

\item{max_degree}{A \code{list} with input indexes for interaction of variables, or a \code{numeric} specifying the maximum max_degree of interaction.}

\item{compl_cost}{A \code{numeric} value specifying the minimum percentage of improvement over the best 1-max_degree basis function to incorporate a higher max_degree basis function.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance. Options are: \code{"mae"}, \code{"mape"}, \code{"mse"}, \code{"rmse"}, \code{"nrmse1"} or \code{"nrmse2"}.}

\item{nterms}{A positive \code{integer} specifying the maximum number of terms created before pruning.}

\item{err_red}{A \code{numeric} value specifying the minimum reduced error rate for the addition of a new pair of 1-max_degree basis functions.}

\item{minspan}{A \code{numeric} value specifying the minimum number of observations between two adjacent knots. Options are: \code{"-2"}, \code{"-1"} or \code{"m"}.}

\item{endspan}{A \code{numeric} value specifying the minimum number of observations before the first and after the final knot. Options are: \code{"-2"}, \code{"-1"} or \code{"m"}.}

\item{kn_grid}{Grid of knots to perform RF-ACES. It can be: \code{-1} or a \code{list}.}
}
\value{
This function return error messages if hyperparameters are incorrectly specified.
}
\description{
This function displays error messages if hyperparameters are incorrectly specified in RF-ACES functions aimed at estimating production functions.
}
