% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_errors.R
\name{display_errors_scores}
\alias{display_errors_scores}
\title{Argument Validation for Efficiency Score Computation}
\usage{
display_errors_scores(data, x, y, object, method, measure, returns, direction)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the DMUs to be evaluated.}

\item{x}{\code{numeric} vector of column indexes for input variables in \code{data}.}

\item{y}{\code{numeric} vector of column indexes for output variables in \code{data}.}

\item{object}{An \code{aces} or \code{rf_aces} object.}

\item{method}{Prediction method (\code{"aces"}, \code{"aces_cubic"}, \code{"rf_aces"}, etc.) used to obtain virtual outputs.}

\item{measure}{Mathematical programming model (\code{"rad_out"}, \code{"ddf"}, \code{"rsl_out"}, etc.).}

\item{returns}{Returns to scale assumption: \code{"constant"} (CRS) or \code{"variable"} (VRS).}

\item{direction}{Projection vector for DDF.}
}
\value{
None. The function is used for its side effect of generating \code{stop()} messages upon validation failure.
}
\description{
Validates input data and hyperparameters for efficiency estimation functions in the ACES framework. It halts execution with a descriptive error message if any parameter is misspecified.
}
