% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_errors.R
\name{display_errors_scores}
\alias{display_errors_scores}
\title{Error Messaging in Adaptive Constrained Enveloping Splines (ACES) functions II}
\usage{
display_errors_scores(
  data,
  x,
  y,
  object,
  measure,
  returns,
  direction,
  weights,
  digits
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{object}{An \code{aces} object.}

\item{measure}{Mathematical programming model to compute the efficiency scores.}

\item{returns}{Type of returns to scale for computing the efficiency scores.}

\item{direction}{Vector direction for DMU projection in Directional Distance Function when computing the efficiency scores.}

\item{weights}{Weights for the additive model.}

\item{digits}{Number of digits to round efficiency scores.}
}
\value{
This function returns error messages if hyperparameters are incorrectly specified.
}
\description{
This function displays error messages if hyperparameters are incorrectly specified in ACES functions aimed at computing efficiency scores.
}
