% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R, R/display_errors.R
\name{err_metric}
\alias{err_metric}
\title{Error Metric for Model Evaluation.}
\usage{
err_metric(y_obs, y_hat, metric)

err_metric(y_obs, y_hat, metric)
}
\arguments{
\item{y_obs}{Vector of observed data.}

\item{y_hat}{Vector of predicted values.}

\item{metric}{Lack-of-fit criterion to evaluate the model performance:
\itemize{
\item{\code{mae}}: Mean Absolute Error
\item{\code{mape}}: Mean Absolute Percentage Error
\item{\code{mse}}: Mean Squared Error
\item{\code{rmse}}: Root Mean Squared Error
\item{\code{nrmse1}}: Normalized Root Mean Squared Error (using mean)
\item{\code{nrmse2}}: Normalized Root Mean Squared Error (using range)
}}
}
\value{
The calculated error metric.

The calculated error metric.
}
\description{
Computes an error metric for model evaluation based on observed and predicted values.

Computes an error metric for model evaluation based on observed and predicted values.
}
