% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_linear.R
\name{estimate_coefficients}
\alias{estimate_coefficients}
\title{Estimate Coefficients for Adaptive Constrained Enveloping Splines (ACES)}
\usage{
estimate_coefficients(
  model_type,
  B,
  y_obs,
  dea_scores,
  it_list,
  Bp_list,
  shape
)
}
\arguments{
\item{model_type}{A \code{character} string specifying the nature of the production frontier.}

\item{B}{A \code{matrix} of linear basis functions derived from input variables.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{dea_scores}{A \code{matrix} containing DEA-VRS efficiency scores, calculated using an output-oriented radial model. For models with multiple outputs, each column corresponds to the scores for one specific output.}

\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity.}
}
\value{
A \code{vector} containing the estimated coefficients for the ACES model.
}
\description{
This function estimates a vector of coefficients for the Adaptive Constrained Enveloping Splines (ACES) model. These coefficients enforce the desired shape properties, such as monotonicity and concavity to fit the data within the ACES framework.
}
