% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients.R
\name{estimate_coefficients}
\alias{estimate_coefficients}
\title{Estimate Coefficients during Forward and Backward Algorithms.}
\usage{
estimate_coefficients(B, y, it_list, Bp_list, monotonicity, concavity, x0_y0)
}
\arguments{
\item{B}{A \code{matrix} of linear basis functions.}

\item{y}{A \code{matrix} of output data.}

\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}
}
\value{
A \code{vector} of estimated coefficients.
}
\description{
This function solves a Linear Programming problem to obtain a set of coefficients (that impose concavity and monotonicity) in the Adaptive Constrained Enveloping Splines estimator.
}
