% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients_smooth.R
\name{estimate_coefficients_smoothed}
\alias{estimate_coefficients_smoothed}
\title{Estimate Coefficients for Smooth Adaptive Constrained Enveloping Splines (ACES) Fitting}
\usage{
estimate_coefficients_smoothed(
  model_type,
  B,
  y_obs,
  dea_eff,
  n_pair,
  n_lsub,
  shape
)
}
\arguments{
\item{model_type}{A \code{character} string specifying the nature of the production frontier that the function will estimate.}

\item{B}{A \code{matrix} of cubic or quintic basis functions.}

\item{y_obs}{A \code{matrix} of the observed output data.}

\item{dea_eff}{An indicator vector with ones for efficient DMUs and zeros for inefficient DMUs.}

\item{n_pair}{An \code{integer} specifying the number of coefficients associated with the paired basis functions.}

\item{n_lsub}{An \code{integer} specifying the number of coefficients associated with the unpaired left-side basis functions.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity and/or passing through the origin.}
}
\value{
A \code{vector} with the estimated coefficients.
}
\description{
This function solves a Linear Programming problem to obtain a set of coefficients that impose concavity and non-decreasing monotonicity in the Smooth Adaptive Constrained Enveloping Splines estimator.
}
