% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients.R
\name{estimate_coefficients_smoothed}
\alias{estimate_coefficients_smoothed}
\title{Estimate Coefficients during Shape-Restricted Smoothing Procedure.}
\usage{
estimate_coefficients_smoothed(B, y, h, r, monotonicity, concavity, x0_y0)
}
\arguments{
\item{B}{A \code{matrix} of cubic / quintic basis functions.}

\item{y}{The output \code{vector} in data.}

\item{h}{Number of coefficients associated with the paired basis functions.}

\item{r}{Number of coefficients associated with the unpaired left-side basis functions.}

\item{monotonicity}{\code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{\code{logical} indicating whether to enforce the constraint of concavity in the estimator.}

\item{x0_y0}{\code{logical} indicating if f(0) = 0.}
}
\value{
A \code{vector} with the estimated coefficients.
}
\description{
This function solves a Linear Programming problem to obtain a set of coefficients (that impose concavity and monotonicity) in the Smoothing Adaptive Constrained Enveloping Splines estimator.
}
