% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackwardAlgorithm.R
\name{EstimCoeffsBackward}
\alias{EstimCoeffsBackward}
\title{Estimate Coefficients in Additive Adaptive Frontier Splines during Backward Procedure.}
\usage{
EstimCoeffsBackward(B, y, h, r)
}
\arguments{
\item{B}{\code{matrix} of basis functions.}

\item{y}{Output \code{vector} in data.}

\item{h}{Number of coefficients associated with paired basis functions.}

\item{r}{Number of coefficients associated with unpaired left-side basis functions.}
}
\value{
\code{vector} with the coefficients estimated.
}
\description{
This function solves a Linear Programming Problem to obtain a set of coefficients.
}
