% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACES.R
\name{generate_technology}
\alias{generate_technology}
\title{Generate a Technology Set}
\usage{
generate_technology(
  tech_xmat,
  tech_ymat1,
  tech_ymat2,
  error_type,
  table_scores,
  ptto
)
}
\arguments{
\item{tech_xmat}{A \code{matrix} representing the input data.}

\item{tech_ymat1}{A \code{matrix} representing the output data.}

\item{tech_ymat2}{A \code{matrix} representing the output data generated from an ACES model.}

\item{error_type}{A \code{character} string specifying the error structure when fitting the model.}

\item{table_scores}{A \code{matrix} containing radial output scores using all outputs and using just each individual output.}

\item{ptto}{A \code{logical} indicating if (0, 0) should be included in the technology.}
}
\value{
A \code{matrix} representing the technology set.
}
\description{
This function generates a technology set based on input data and an ACES model output.
The technology set represents the feasible combinations of inputs and outputs.
}
