% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predictions.R
\name{get_targets}
\alias{get_targets}
\title{Compute Efficient Targets using Adaptive Constrained Enveloping Splines (ACES)}
\usage{
get_targets(
  eval_data,
  x,
  y,
  relevant = FALSE,
  object,
  method = "aces",
  measure = "rad_out",
  returns = "variable",
  direction = NULL
)
}
\arguments{
\item{eval_data}{A \code{data.frame} or a \code{matrix} containing the DMUs to be evaluated.}

\item{x}{Column indexes of input variables in \code{eval_data}.}

\item{y}{Column indexes of output variables in \code{eval_data}.}

\item{relevant}{A \code{logical} indicating if only relevant variables should be included in the technology definition.}

\item{object}{An \code{aces} object.}

\item{method}{Model prediction method used to compute predictions of inputs and obtain a new vector of outputs for \code{eval_data}:
\itemize{
\item{\code{"aces_forward"}}: Forward Adaptive Constrained Enveloping Splines.
\item{\code{"aces"}}: Adaptive Constrained Enveloping Splines.
\item{\code{"aces_cubic"}}: Cubic Smooth Adaptive Constrained Enveloping Splines.
\item{\code{"aces_quintic"}}: Quintic Smooth Adaptive Constrained Enveloping Splines.
}}

\item{measure}{Mathematical programming model to calculate scores:
\itemize{
\item{\code{rad_out}} Output-oriented radial measure proposed by \insertCite{banker1984;textual}{aces}.
\item{\code{rad_inp}} Input-oriented radial measure proposed by \insertCite{banker1984;textual}{aces}.
\item{\code{ddf}}     Directional distance function proposed by \insertCite{chambers1998;textual}{aces}.
\item{\code{rsl_out}} Output-oriented Russell measure proposed by \insertCite{fare1978;textual}{aces}.
\item{\code{rsl_inp}} Input-oriented Russell measure proposed by \insertCite{fare1978;textual}{aces}.
\item{\code{wam_mip}} Measure of Inefficiency Proportions proposed by \insertCite{cooper1999;textual}{aces}.
\item{\code{wam_nor}} Normalized Weighted Additive Model proposed by \insertCite{lovell1995;textual}{aces}.
\item{\code{wam_ram}} Range Adjusted Measure proposed by \insertCite{cooper1999;textual}{aces}.
\item{\code{wam_bam}} Bounded Adjusted Measure proposed by \insertCite{cooper2011;textual}{aces}.
}}

\item{returns}{Type of returns to scale:
\itemize{
\item{\code{"constant"}} Constant returns to scale.
\item{\code{"variable"}} Variable returns to scale (default).
}}

\item{direction}{Direction of the vector to project on the frontier. Only applied if \code{measure = "ddf"}. A \code{matrix} or \code{data.frame} with \code{n} rows (number of DMUs to be evaluated) and \code{nX + nY} columns, containing the direction of the input variables followed by the direction of the output variables in the same order as they appear in the data.}
}
\value{
A \code{data.frame} with input and output targets computed through an Adaptive Constrained Enveloping Splines model.
}
\description{
Projects a set of Decision Making Units (DMUs) onto the estimated production frontier. This function calculates the efficient input/output targets required for each DMU to achieve full efficiency, based on the specified ACES model and distance measure.
}
\references{
\insertRef{espana2024}{aces} \cr \cr
\insertRef{banker1984}{aces} \cr \cr
\insertRef{chambers1998}{aces} \cr \cr
\insertRef{fare1978}{aces} \cr \cr
\insertRef{cooper1999}{aces} \cr \cr
\insertRef{lovell1995}{aces} \cr \cr
\insertRef{cooper2011}{aces}
}
