% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technology.R
\name{get_technology}
\alias{get_technology}
\title{Get Technology Set}
\usage{
get_technology(object, method = "aces")
}
\arguments{
\item{object}{An \code{aces} object.}

\item{method}{Model for prediction:
\itemize{
\item{\code{"aces_forward"}}: Forward Adaptive Constrained Enveloping Splines.
\item{\code{"aces"}}: Adaptive Constrained Enveloping Splines.
\item{\code{"aces_cubic"}}: Cubic Smoothed Adaptive Constrained Enveloping Splines.
\item{\code{"aces_quintic"}}: Quintic Smoothed Adaptive Constrained Enveloping Splines.
}}
}
\value{
A matrix containing the inputs (xmat) and the estimated output (ymat).
}
\description{
Retrieves the matrix constituting the estimated technology for a specific method from an ACES object.
}
