% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAFS.R
\name{MAFS}
\alias{MAFS}
\title{Multivariate Adaptive Frontier Splines}
\usage{
MAFS(
  data,
  x,
  y,
  nterms,
  Kp = 1,
  d = 2,
  err.red = 0.01,
  minspan = -2,
  endspan = -1,
  knotsGrid = -1,
  linpreds = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{nterms}{Maximum number of terms created by the forward algorithm (before pruning).}

\item{Kp}{Maximum degree of interaction allowed. Default is \code{1}.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot. Default is \code{2}. If it is set to \code{-1}, \code{GCV = RSS / n}.}

\item{err.red}{Minimum reduced error rate for the addition of two new basis functions. Default is \code{0.01}.}

\item{minspan}{Minimum number of observations between knots.
\itemize{
\item{\code{minspan = -2}} Minspan computed as in Zhang (1994).
\item{\code{minspan = -1}} Minspan computed as in Friedman's MARS paper section 3.8 with alpha = 0.05.
\item{\code{minspan =  m}} \code{m} observations between knots.
}}

\item{endspan}{Minimum number of observations before the first and after the final knot.
\itemize{
\item{\code{endspan = -2}} Endspan computed as in Zhang (1994).
\item{\code{endspan = -1}} Endspan computed as in Friedman's MARS paper section 3.8 with alpha = 0.05.
\item{\code{endspan = +m}} \code{m} observations between knots.
}}

\item{knotsGrid}{Grid of knots to perform MAFS:
\itemize{
\item{\code{knotsGrid = -1}} The original approach based on the observed data is used.
\item{\code{knotsGrid =  0}} A grid of equidistant knots is created from the scattering of the data.
\item{\code{knotsGrid = +p}} \code{p} percentiles are computed to perform the grid of knots.
}}

\item{linpreds}{\code{logical}. If \code{TRUE}, predictors can enter linearly.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
A \code{MAFS} object.
}
\description{
This function estimates a production frontier satisfying some classical production theory axioms, such as monotonicity and concavity, which is based upon the adaptation of the machine learning technique known as Multivariate Adaptive Regression Splines (MARS).
}
