% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAFS.R
\name{MAFS_object}
\alias{MAFS_object}
\title{Create a MAFS object}
\usage{
MAFS_object(
  data,
  x,
  y,
  nterms,
  Kp,
  d,
  err.red,
  na.rm,
  Forward.MAFS,
  MAFS,
  QSmooth.MAFS,
  CSmooth.MAFS
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in \code{data}.}

\item{y}{Column output indexes in \code{data}.}

\item{nterms}{Maximum number of terms created by the forward algorithm (before pruning).}

\item{Kp}{Maximum degree of interaction allowed. Default is \code{1}.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot. Default is \code{2}. If set to \code{-1}, \code{GCV = RSS / n}.}

\item{err.red}{Minimun reduced error rate for the addition of two new basis functions. Default is \code{0.01}.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.}

\item{Forward.MAFS}{\code{list} containing the Forward Multivariate Adaptive Frontier Splines.}

\item{MAFS}{\code{list} containing the Multivariate Adaptive Frontier Splines model.}

\item{QSmooth.MAFS}{\code{list} containing the Smooth Multivariate Adaptive Frontier Splines model (quintic version).}

\item{CSmooth.MAFS}{\code{list} containing the Smooth Multivariate Adaptive Frontier Splines model (cubic version).}
}
\value{
A \code{MAFS} object.
}
\description{
This function saves information about the Multivariate Adapative Frontier Splines model.
}
