% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coefficients.R
\name{monotonocity_matrix}
\alias{monotonocity_matrix}
\title{Build Constraint Matrix for Enforcing Non-decreasing Monotonicity}
\usage{
monotonocity_matrix(it_list, Bp_list, N)
}
\arguments{
\item{it_list}{A \code{list} containing the set of intervals by input.}

\item{Bp_list}{A \code{list} containing the set of basis functions by input.}

\item{N}{Sample size.}
}
\value{
A \code{matrix} representing the constraint matrix to enforce non-decreasing monotonicity on the estimated function.
}
\description{
This function constructs the constraint matrix (often denoted as matrix "A") to enforce non-decreasing monotonicity in an estimated function. The constraint matrix enforces concavity by ensuring that the coefficients of the basis functions satisfy certain conditions.
}
