% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.aces}
\alias{predict.aces}
\title{Model Prediction for Adaptive Constrained Enveloping Splines (ACES).}
\usage{
\method{predict}{aces}(object, newdata, x, method = "aces")
}
\arguments{
\item{object}{An \code{aces} object.}

\item{newdata}{A \code{data.frame} containing the input and netput variables to predict on.}

\item{x}{Input indexes in \code{newdata}.}

\item{method}{Model for prediction.
\itemize{
\item{\code{"aces_forward"}}: Forward Adaptive Constrained Enveloping Splines model.
\item{\code{"aces"}}: Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_cubic"}}: Cubic Smoothed Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_quintic"}}: Quintic Smoothed Adaptive Constrained Enveloping Splines model.
}}
}
\value{
A \code{data.frame} with the predicted values through the Adaptive Constrained Enveloping Splines model.
}
\description{
This function predicts the expected output by an \code{aces} object.
}
