% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.rf_aces}
\alias{predict.rf_aces}
\title{Model Prediction for Random Forest Adaptive Constrained Enveloping Splines (RF-ACES)}
\usage{
\method{predict}{rf_aces}(object, newdata, x, method = "rf_aces", stochastic_pred = FALSE)
}
\arguments{
\item{object}{A \code{rf_aces} object.}

\item{newdata}{A \code{data.frame} containing the input and netput variables to predict on.}

\item{x}{Input indexes in \code{newdata}.}

\item{method}{Model for prediction:
\itemize{
\item{\code{"rf_aces"}}: Random Forest Adaptive Constrained Enveloping Splines model.
\item{\code{"rf_aces_cubic"}}: Random Forest Cubic Smoothed Adaptive Constrained Enveloping Splines model.
\item{\code{"rf_aces_quintic"}}: Random Forest Quintic Smoothed Adaptive Constrained Enveloping Splines model.
}}

\item{stochastic_pred}{This parameter should be kept as \code{FALSE} (default) if an additive model is estimated. If a stochastic model is estimated, one of the following elements should be specified:
\itemize{
  \item{"avg"}: average-practice production function.
  \item{"mom"}: moments-estimation production function.
  \item{"pse"}: pseudolikelihood-estimation production function.
}}
}
\value{
A \code{data.frame} with the predicted values through the Random Forest Adaptive Constrained Enveloping Splines model.
}
\description{
This function predicts the expected output by an \code{rf_aces} object.
}
