% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict.rf_aces}
\alias{predict.rf_aces}
\title{Model Prediction for Random Forest Adaptive Constrained Enveloping Splines (RF-ACES)}
\usage{
\method{predict}{rf_aces}(object, newdata, x, method = "aces")
}
\arguments{
\item{object}{A \code{rf_aces} object.}

\item{newdata}{A \code{data.frame} containing the input and netput variables to predict on.}

\item{x}{Input indexes in \code{newdata}.}

\item{method}{Model for prediction:
\itemize{
\item{\code{"aces_forward"}}: Random Forest Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_cubic"}}: Random Forest Cubic Smoothed Adaptive Constrained Enveloping Splines model.
\item{\code{"aces_quintic"}}: Random Forest Quintic Smoothed Adaptive Constrained Enveloping Splines model.
}}
}
\value{
A \code{data.frame} with the predicted values through the Random Forest Adaptive Constrained Enveloping Splines model.
}
\description{
This function predicts the expected output by an \code{rf_aces} object.
}
