% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{quintic_aces}
\alias{quintic_aces}
\title{Fit a Smooth Adaptive Constrained Enveloping Splines using Quintic Functions}
\usage{
quintic_aces(
  data,
  x,
  y,
  dea_eff,
  model_type,
  metric,
  shape,
  kn_grid,
  kn_side_loc,
  d,
  wq
)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{dea_eff}{Indicator vector with 1s for efficient DMUs and 0s for inefficient DMUs.}

\item{model_type}{A \code{character} string specifying the nature of the production frontier that the function will estimate.}

\item{metric}{A \code{character} string specifying the lack-of-fit criterion to evaluate the model performance.}

\item{shape}{A \code{list} indicating whether to impose monotonicity and/or concavity and/or passing through the origin.}

\item{kn_grid}{A \code{data.frame} containing the set of knots from the backward step for smoothing.}

\item{kn_side_loc}{A \code{list} with the side knot locations for each input dimension.}

\item{d}{Generalized Cross Validation (GCV) penalty per knot.}

\item{wq}{Let `p` be the distance between the central knot and the right-side knot, and `v` be the distance between the central knot and the left-side knot during the smoothing procedure. A \code{numeric} value used for quintic smoothing \insertCite{chen1999}{aces}. This parameter is defined as `p / v` and must be set between 8/7 and 1.5. If a \code{vector} is entered, the \code{wc} value that most reduced the lack-of-fit criterion is selected.}
}
\value{
A \code{list} containing information relative to the Smooth Adaptive Constrained Enveloping Splines through quintic functions.
}
\description{
This function performs smoothing on the Adaptive Constrained Enveloping Splines predictor while enforcing continuous second derivatives.
}
\references{
\insertRef{chen1999}{aces}
}
