% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_ACES.R
\name{rf_aces}
\alias{rf_aces}
\title{Fit a Random Forest Adaptive Constrained Enveloping Splines (RF-ACES) model.}
\usage{
rf_aces(
  data,
  x,
  y,
  quick_aces = TRUE,
  model_type = "env",
  error_type = "add",
  mul_BF = list(max_degree = 1, inter_cost = 0.05),
  metric = "mse",
  shape = list(mono = TRUE, conc = TRUE, ptto = FALSE),
  learners = 100,
  nvars = length(x)/3,
  sample_size = nrow(data),
  nterms = 50,
  err_red = 0.01,
  kn_grid = -1,
  minspan = -1,
  endspan = -1
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{quick_aces}{A \code{logical} indicating if the fast version of ACES should be employed.}

\item{error_type}{A \code{character} string specifying the error structure to use. Options are:
\itemize{
  \item{\code{"add"}}: Additive error structure.
  \item{\code{"mul"}}: Multiplicative error structure.
}}

\item{mul_BF}{A \code{list} specifying the maximum degree of basis functions (BFs) and the cost of introducing a higher-degree BF. Items include:
\itemize{
\item{\code{max_degree}}: A \code{list} with input indexes for interaction of variables, or a \code{numeric} specifying the maximum degree of interaction. BFs products are constrained to contain factors involving distinct variables to ensure interpretability and avoid multicollinearity.
\item{\code{inter_cost}}: A \code{numeric} specifying the minimum percentage improvement over the best 1-degree BF to incorporate a higher degree BF. Default is \code{0.05}.
}}

\item{metric}{A \code{list} specifying the lack-of-fit criterion to evaluate the model performance. Options are:
\itemize{
\item{\code{"mae"}}: Mean Absolute Error.
\item{\code{"mape"}}: Mean Absolute Percentage Error.
\item{\code{"mse"}}: Mean Squared Error.
\item{\code{"msle"}}: Mean Squared Logarithmic Error.
\item{\code{"rmse"}}: Root Mean Squared Error.
\item{\code{"nrmse1"}}: Normalized Root Mean Squared Error (using mean).
\item{\code{"nrmse2"}}: Normalized Root Mean Squared Error (using range).
}}

\item{shape}{A \code{list} specifying shape constraints for the estimator. Items include:
\itemize{
  \item{\code{mono}}: A \code{logical} indicating if non-decreasing monotonicity should be enforced.
  \item{\code{conc}}: A \code{logical} indicating if concavity should be enforced.
  \item{\code{ptto}}: A \code{logical} indicating if the estimator should satisfy \code{f(0) = 0}.
}}

\item{learners}{An \code{integer} indicating the number of models for bagging.}

\item{nvars}{An \code{integer} indicating the number of variables randomly chosen at each split in RF-ACES.}

\item{sample_size}{An \code{integer} indicating the number of samples to draw from \code{data} to train each base estimator.}

\item{nterms}{A positive \code{integer} specifying the maximum number of terms created during the forward step. Default is \code{50}.}

\item{err_red}{A \code{numeric} value specifying the minimum reduced error rate for the addition of a new pair of 1-degree basis functions. Default is \code{0.01}.}

\item{kn_grid}{Specifies the grid of knots to be used in the ACES algorithm. Accepts two options:
\itemize{
   \item{\code{-1}} (default): Uses the original approach by \insertCite{friedman1991;textual}{aces}, where the knots are automatically selected based on the observed data.
   \item{\code{list}}: A user-defined grid of knots for each variable. This must be a \code{list} where each element corresponds to a vector of knots for a specific variable (e.g., the first element of the \code{list} contains the knot values for the first variable, the second element contains the knots for the second variable, and so on). This option allows for greater control over the knot placement when customizing the estimation process.

}}

\item{minspan}{A \code{numeric} specifying the minimum number of observations between two adjacent knots. Options are:
\itemize{
\item{\code{minspan = -2}}: Computed as in \insertCite{zhang1994;textual}{aces}.
\item{\code{minspan = -1}}: Computed as in \insertCite{friedman1991;textual}{aces}.
\item{\code{minspan = +m}}: A user-specified positive integer.
}}

\item{endspan}{A \code{numeric} specifying the minimum number of observations before the first and after the final knot. Options are:
\itemize{
\item{\code{endspan = -2}}: Computed as in \insertCite{zhang1994;textual}{aces}.
\item{\code{endspan = -1}}: Computed as in \insertCite{friedman1991;textual}{aces}.
\item{\code{endspan = +m}}: A user-specified positive integer.
}}
}
\value{
An \code{rf_aces} object.
}
\description{
This function estimates a deterministic production frontier that adheres to classical production theory axioms, such as monotonicity and concavity. The estimation is based on adaptations of the Multivariate Adaptive Regression Splines (MARS) technique, developed by \insertCite{friedman1991;textual}{aces} and Random Forest, introduced by \insertCite{breiman2001;textual}{aces}. For comprehensive details on the methodology and implementation, please refer to \insertCite{espana2024;textual}{aces} and \insertCite{espana2024rf;textual}{aces}.
}
\details{
See ...
}
\references{
\insertRef{espana2024rf}{aces} \cr \cr
\insertRef{espana2024}{aces} \cr \cr
\insertRef{friedman1991}{aces} \cr \cr
\insertRef{breiman2001}{aces} \cr \cr
\insertRef{zhang1994}{aces} \cr
}
