% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_ACES.R
\name{rf_aces_p_predict}
\alias{rf_aces_p_predict}
\title{Percentile-Based Prediction for Random Forest Adaptive Constrained Enveloping Splines (RF-ACES).}
\usage{
rf_aces_p_predict(object, newdata, x, y, p, method)
}
\arguments{
\item{object}{A \code{rf_aces} object.}

\item{newdata}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{p}{A numeric value between 0 and 1 specifying the conditional percentile to be used for prediction.}

\item{method}{Model for prediction:
\itemize{
\item{\code{"rf_aces"}}: Random Forest Adaptive Constrained Enveloping Splines.
\item{\code{"rf_aces_cubic"}}: Random Forest Cubic Smoothed Adaptive Constrained Enveloping Splines.
\item{\code{"rf_aces_quintic"}}: Random Forest Quintic Smoothed Adaptive Constrained Enveloping Splines.
}}
}
\value{
A numeric vector of predicted output values for each observation, based on the specified percentile.
}
\description{
Predicts the optimal output level under the RF-ACES model using a specified percentile of the conditional distribution, rather than the mean. As in DEA, inputs are held fixed, and the output is predicted via Random Forests by selecting a given percentile (e.g., the 5th percentile) of the conditional distribution. This approach enables modeling conservative or optimistic production frontiers by adjusting the percentile parameter.
}
