% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_ACES.R
\name{rf_aces_varimp}
\alias{rf_aces_varimp}
\title{Compute Variable Importance for Random Forest Adaptive Constrained Enveloping Splines (RF-ACES).}
\usage{
rf_aces_varimp(data, x, y, object, repeats = 1)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column indexes of input variables in \code{data}.}

\item{y}{Column indexes of output variables in \code{data}.}

\item{object}{A \code{rf_aces} object.}

\item{repeats}{Number of times the variable importance procedure is repeated.}
}
\value{
This function returns a metric of variable importance for each input variable.
}
\description{
Computes a robust measure of variable importance for a fitted Random Forest Adaptive Constrained Enveloping Splines (RF-ACES) model. Importance scores are obtained by evaluating the increase in prediction error when permuting each input variable as in \insertCite{breiman2001;textual}{aces}.
}
\references{
\insertRef{breiman2001}{aces} \cr
}
