% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{set_Bmat}
\alias{set_Bmat}
\title{Build (B) Matrix of Basis Functions}
\usage{
set_Bmat(newdata, model, knots, method)
}
\arguments{
\item{newdata}{A \code{data.frame} containing the input and netput variables to predict on.}

\item{model}{A \code{list} with data of a model.}

\item{knots}{Set of knots of the \code{model}.}

\item{method}{Type of model for prediction.}
}
\value{
The (B) Matrix of Basis Functions
}
\description{
This function builds the (B) matrix of basis functions for prediction given a model.
}
