% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_algorithm.R
\name{set_intervals}
\alias{set_intervals}
\title{Generate the Set of Intervals}
\usage{
set_intervals(Bp_list)
}
\arguments{
\item{Bp_list}{A \code{list} containing the set of basis functions by input.}
}
\value{
A \code{list} with the boundaries of the intervals and the IDs of the activated basis functions within each interval.
}
\description{
This function determines the boundaries for the intervals and identifies what basis functions activates on each interval.
}
