% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{set_triplet_knots}
\alias{set_triplet_knots}
\title{Generate a Suitable Triplet of Knots}
\usage{
set_triplet_knots(knots, w, smoothing, mono, conc)
}
\arguments{
\item{knots}{A \code{vector} with the initial locations for central and side knots.}

\item{w}{Hyperparameter for side knot distances in the cubic and quintic smoothing procedures.}

\item{smoothing}{Type of smoothing:
\itemize{
\item{\code{"cubic"}}: cubic smoothing procedure.
\item{\code{"quintic"}}: quintic smoothing procedure.
}}

\item{monotonicity}{A \code{logical} indicating whether to enforce the constraint of non-decreasing monotonicity in the estimator.}

\item{concavity}{A \code{logical} indicating whether to enforce the constraint of concavity in the estimator.}
}
\value{
A \code{vector} with a suitable triplet of knots.
}
\description{
This function generates a suitable triplet of knots for the smoothing procedure depending on the shape-constraints.
}
