% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_algorithm.R
\name{side_knot_location}
\alias{side_knot_location}
\title{Generate Side Knot Locations}
\usage{
side_knot_location(data, nX, knots)
}
\arguments{
\item{data}{A \code{matrix} containing the variables in the model.}

\item{nX}{Number of inputs in \code{data}.}

\item{knots}{A \code{data.frame} containing the knots from the backward step.}
}
\value{
A \code{list} with the locations of the central and the side knots in each dimension.
}
\description{
This function creates the input space for locating side knots in the smoothing procedure.
}
