% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{update_Bp_list}
\alias{update_Bp_list}
\title{Update Bp_list}
\usage{
update_Bp_list(basis, nX)
}
\arguments{
\item{basis}{A \code{list} containing the set of basis functions.}

\item{nX}{Number of inputs.}
}
\value{
A \code{list} with the basis functions updated.
}
\description{
This function updates the Bp_list during the backward algorithm.
}
