% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_algorithm.R
\name{update_knot_list}
\alias{update_knot_list}
\title{Update knot_list}
\usage{
update_knot_list(Bp_list)
}
\arguments{
\item{Bp_list}{A \code{list} containing the set of basis functions by input.}
}
\value{
A \code{list} with the knots updated.
}
\description{
This function updates the knot_list during the backward algorithm.
}
