% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_scores.R
\name{wam}
\alias{wam}
\title{The Weighted Additive Model}
\usage{
wam(
  tech_xmat,
  tech_ymat,
  eval_xmat,
  eval_ymat,
  weights,
  convexity,
  returns,
  type
)
}
\arguments{
\item{tech_xmat}{A \code{matrix} containing the observed input variables used to define the technology.}

\item{tech_ymat}{A \code{matrix} containing the observed output variables used to define the technology.}

\item{eval_xmat}{A \code{matrix} containing the input data of the Decision-Making Units (DMUs) to be evaluated.}

\item{eval_ymat}{A \code{matrix} containing the output data of the Decision-Making Units (DMUs) to be evaluated.}

\item{weights}{Weights for the additive model:
\itemize{
\item{\code{"wam.one"}} Weighted Additive Model.
\item{\code{"wam.mip"}} Measure of Inefficiency Proportions.
\item{\code{"wam.nor"}} Normalized Weighted Additive Model.
\item{\code{"wam.ram"}} Range Adjusted Measure.
\item{\code{"wam.bam"}} Bounded Adjusted Measure.
}}

\item{convexity}{A \code{logical} value indicating if a convex technology is assumed.}

\item{returns}{Type of returns to scale. Either \code{"constant"} (CRS) or \code{"variable"} (VRS).}

\item{type}{A \code{character} string specifying the return value. Options are \code{"objective"} to return the optimal value of the objective function, and \code{"variables"} to return the optimal values of the decision variables related to efficiency measure.}
}
\value{
A \code{vector} of \code{"numeric"} scores computed through the Weighted Additive Model.
}
\description{
This function computes efficiency scores through a Weighted Additive Model.
}
