% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_scores.R
\name{wam}
\alias{wam}
\title{The Weighted Additive Model}
\usage{
wam(tech_xmat, tech_ymat, eval_xmat, eval_ymat, weights, convexity, returns)
}
\arguments{
\item{tech_xmat}{A \code{matrix} containing the observed input variables used to define the technology.}

\item{tech_ymat}{A \code{matrix} containing the observed output variables used to define the technology.}

\item{eval_xmat}{A \code{matrix} containing the input data of the Decision-Making Units (DMUs) to be evaluated.}

\item{eval_ymat}{A \code{matrix} containing the output data of the Decision-Making Units (DMUs) to be evaluated.}

\item{weights}{Weights for the additive model:
\itemize{
\item{\code{"WAM"}} Weighted Additive Model.
\item{\code{"MIP"}} Measure of Inefficiency Proportions.
\item{\code{"NOR"}} Normalized Weighted Additive Model.
\item{\code{"RAM"}} Range Adjusted Measure.
\item{\code{"BAM"}} Bounded Adjusted Measure.
}}

\item{convexity}{A \code{logical} value indicating if a convex technology is assumed.}

\item{returns}{Type of returns to scale.}
}
\value{
A \code{vector} of \code{"numeric"} scores computed through the Weighted Additive Model.
}
\description{
This function computes efficiency scores through a Weighted Additive Model.
}
